Manually defined split

Manually defined split

ManualSplit(
  levels,
  label,
  name = "manual",
  extra_args = list(),
  indent_mod = 0L,
  cindent_mod = 0L,
  cvar = "",
  cextra_args = list(),
  label_pos = "visible",
  page_prefix = NA_character_
)

Arguments

levels

character. Levels of the split (ie the children of the manual split)

label

character(1). A label (not to be confused with the name) for the object/structure.

name

character(1). Name of the split/table/row being creatted. Defaults to same as the corresponding llabel, but is not required to be.

extra_args

list. Extra arguments to be passed to the tabulation function. Element position in thte list corresponds to the children of this split. Named elements in the child-specific lists are ignored if they do not match a formal argument of the ttabulation function.

indent_mod

numeric. Modifier for the default indent position for the structure created by this function(subtable, content table, or row) and all of that structure's children. Defaults to 0, which corresponds to the unmodified default behavior.

cindent_mod

numeric(1). The indent modifier for the content tables generated by this split.

cvar

character(1). The variable, if any, which the content function should accept. Defaults to NA.

cextra_args

list. Extra arguments to be passed to the content function when tabulating row group summaries.

label_pos

character(1). Location the variable label should be displayed, Accepts hidden (default for non-analyze row splits), visible, topleft, and - for analyze splits only - default. For analyze calls, default indicates that the variable should be visible if and only if multiple variables are analyzed at the same level of nesting.

page_prefix

character(1). Prefix, to be appended with the split value, when forcing pagination between the children of this split/table

Value

A ManualSplit object.

Author

Gabriel Becker