Generate Rows Analyzing Variables Across Columns

Adding /analyzed variables/ to our table layout defines the primary tabulation to be performed. We do this by adding calls to add_analyzed_vars and/or add_analyzed_colvars into our layout pipeline. As with adding further splitting, the tabulation will occur at the current/next level of nesting by default.

add_analyzed_vars(lyt, var, lbl = var, afun, fmt = NULL,
  rowlabs = "", newtoplev = FALSE, inclNAs = FALSE)

Arguments

lyt
layout object pre-data used for tabulation

var
string, variable name

lbl
string, label, often variable label

afun
function. Analysis function.

fmt
format string for cells

newtoplev
boolean, Add this as a new top-level split (defining a new subtable directly under root). Defaults to FALSE

inclNAs
boolean. Should observations with NA in the var variable(s) be included when performing this analysis. Defaults to FALSE

Examples

library(magrittr)

l <- NULL %>% add_colby_varlevels("ARM") %>%
    add_analyzed_vars("AGE", afun = lstwrap(summary) , fmt = "xx.xx")
l

#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ARM (lvls) 
#> 
#> Row-Split Structure:
#> AGE (** analysis **) 
#> 

build_table(l, DM)

#>             B: Placebo   C: Combination   A: Drug X
#> ---------------------------------------------------
#> AGE                                                
#>   Min.          21             22            20    
#>   1st Qu.       29             30            29    
#>   Median        32             33            33    
#>   Mean        33.02          34.57          34.91  
#>   3rd Qu.       37             38            39    
#>   Max.          55             53            60