Declaring a column-split based on levels of a variable

Will generate children for each subset of a categorical variable

add_colby_varlevels(lyt, var, lbl = var, valuelblvar = var,
  splfmt = NULL, newtoplev = FALSE)

Arguments

lyt
layout object pre-data used for tabulation

var
string, variable name

lbl
string, label, often variable label

valuelblvar
string, name of variable containing split value display lables

splfmt
format spec. Format associated with this split.

newtoplev
boolean, Add this as a new top-level split (defining a new subtable directly under root). Defaults to FALSE

Examples

library(magrittr)

l <- NULL %>% add_colby_varlevels("ARM")
l

#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ARM (lvls) 
#> 
#> Row-Split Structure:
#> ( () ->  () -> ) () 
#> 

# add an analysis (summary)
l2 <- l %>%
    add_analyzed_vars("AGE", afun = lstwrap(summary) , fmt = "xx.xx")
l2

#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ARM (lvls) 
#> 
#> Row-Split Structure:
#> AGE (** analysis **) 
#> 

build_table(l2, DM)

#>             B: Placebo   C: Combination   A: Drug X
#> ---------------------------------------------------
#> AGE                                                
#>   Min.          21             22            20    
#>   1st Qu.       29             30            29    
#>   Median        32             33            33    
#>   Mean        33.02          34.57          34.91  
#>   3rd Qu.       37             38            39    
#>   Max.          55             53            60    

# By default sequentially adding layouts results in nesting
l3 <- NULL %>% add_colby_varlevels("ARM") %>%
  add_colby_varlevels("SEX") %>%
  add_analyzed_vars("AGE", afun = lstwrap(summary), fmt = "xx.xx")
l3

#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ARM (lvls) -> SEX (lvls) 
#> 
#> Row-Split Structure:
#> AGE (** analysis **) 
#> 

build_table(l3, DM)

#>              B: Placebo      C: Combination       A: Drug X  
#>               M       F        M         F        M       F  
#> -------------------------------------------------------------
#> AGE                                                          
#>   Min.       21      21       25        22       24      20  
#>   1st Qu.    28      29       29        30       31      29  
#>   Median     31      33       32        35       35      32  
#>   Mean      32.1    33.84    34.28     34.89    36.55   33.71
#>   3rd Qu.   35.75    38       38        39      41.5     38  
#>   Max.       47      55       53        53       60      58