Dummy Function with aruments that are used throughout the framework

Dummy Function with aruments that are used throughout the framework

Arguments

lyt
layout object pre-data used for tabulation

var
string, variable name

vars
character vector. Multiple variable names.

lbl
string, label, often variable label

df
dataset

valuelblvar
string, name of variable containing split value display lables

varlbls
character vector. Labels for vars

splfmt
format spec. Format associated with this split.

newtoplev
boolean, Add this as a new top-level split (defining a new subtable directly under root). Defaults to FALSE

spl
A Split object defining a partitioning or analysis/tabulation of the data.

pos
numeric. Which top-level set of nested splits should the new layout feature be added to. Defaults to the current

fmt
format string for cells

cfun
function/NULL. tabulation function for creating content rows

cfmt
format spec. Format for content rows

splfun
function/NULL. custom splitting function

splname
string. Name associiated with this split (for pathing, etc)

splbl
string. Labe associated with this split

afun
function. Analysis function.

defrowlab
character vector. Default row labels for resulting analysis rows.

inclNAs
boolean. Should observations with NA in the var variable(s) be included when performing this analysis. Defaults to FALSE

valorder
character vector. Order that the split children should appear in resulting table.

baseline
character. Value of var to be taken as the baseline/control to be compared against.

compfun
function/string. The comparison function which accepts the analysis function outputs for two different partitions and returns a single value. Defaults to subraction. If a string, taken as the name of a function.

rowlblf
string. An sprintf style format string containing up to one "%s" which takes the current split value and generates the row label for the summary.

tt
TableTree. A TableTree object representing a populated table.

verbose
logical. Should additional information be displayed to the user. Defaults to FALSE.

colwidths
numeric vector. Column widths for use with vertical pagination. Currently ignored.