Convert an rtable object to an shiny.tag html representation of the rtable

The returned html object can be immediately used in shiny and rmarkdown

as_html(x, ...)

Arguments

x
an rtable object

...
arguments passed as attributes to the table html objet

Value

an object of class shiny.tag

Examples

tbl <- rtable(
  col.names = LETTERS[1:3],
  format = "xx",
  rrow("r1", 1,2,3),
  rrow("r2", 4,3,2, indent = 1),
  rrow("r3", indent = 2)
)

#> Error in rtable(col.names = LETTERS[1:3], format = "xx", rrow("r1", 1,     2, 3), rrow("r2", 4, 3, 2, indent = 1), rrow("r3", indent = 2)): not all arguments in ... are of class rrow

as_html(tbl)

#> <table class="table table-condensed table-hover">
#>   <tr>
#>     <th class="rowname text-left"></th>
#>     <th class="text-center">col_1</th>
#>   </tr>
#>   <tr>
#>     <td class="rowname text-left">Sum of Sepal Length</td>
#>     <td class="text-center">876.5 (82.81%)</td>
#>   </tr>
#> </table>