Layout with 1 column and zero rows

Every layout must start with a basic table.

basic_table(
  title = "",
  subtitles = character(),
  main_footer = character(),
  prov_footer = character(),
  show_colcounts = FALSE
)

Arguments

title

character(1). Main title. Ignored for subtables.

subtitles

character. Subtitles. Ignored for subtables.

main_footer

character. Main global (non-referential) footer materials.

prov_footer

character. Provenance-related global footer materials. Generally should not be modified by hand.

show_colcounts

logical(1). Should column counts be displayed in the resulting table when this layout is applied to data

Value

A PreDataTableLayouts object suitable for passing to further layouting functions, and to build_table.

Examples

lyt <- basic_table() %>%
  analyze("AGE", afun = mean)

build_table(lyt, DM)
#>            all obs     
#> ———————————————————————
#> mean   34.2219101123596

lyt <- basic_table(title = "Title of table", subtitles = c("a number", "of subtitles"),
                   main_footer = "test footer",
                   prov_footer = paste("test.R program, executed at", Sys.time())) %>%
  split_cols_by("ARM") %>%
  analyze("AGE", mean)

build_table(lyt, DM)
#> Title of table
#> a number
#> of subtitles
#> 
#> —————————————————————————————————————————————————————————————
#>           A: Drug X          B: Placebo       C: Combination 
#> —————————————————————————————————————————————————————————————
#> mean   34.9090909090909   33.0188679245283   34.5658914728682
#> —————————————————————————————————————————————————————————————
#> 
#> test footer
#> test.R program, executed at 2022-07-26 11:36:13