Adds column to matrix that corresponds to taking all entries (column of all TRUE)

Adds column to matrix that corresponds to taking all entries (column of all TRUE)

by_add_total(col_by, label = "total", n = NULL)

Arguments

col_by
factor or matrix to add column to, e.g. output of col_by_to_matrix

label
label of new column

n
number of rows in data vector, e.g. length(x), useful when col_by can also be NULL

Value

new matrix with one column added

Examples

x <- factor(c("a", "b", "a", "a", "b"))
mat <- col_by_to_matrix(x)
by_add_total(mat, label = "tot")

#>       a     b  tot
#> 1  TRUE FALSE TRUE
#> 2 FALSE  TRUE TRUE
#> 3  TRUE FALSE TRUE
#> 4  TRUE FALSE TRUE
#> 5 FALSE  TRUE TRUE

by_add_total(x, label = "tot")

#>       a     b  tot
#> 1  TRUE FALSE TRUE
#> 2 FALSE  TRUE TRUE
#> 3  TRUE FALSE TRUE
#> 4  TRUE FALSE TRUE
#> 5 FALSE  TRUE TRUE

if (requireNamespace("testthat", quietly = TRUE)) {
  library(dplyr)
  mat %>% by_add_total(label = "tot")
}

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

#>       a     b  tot
#> 1  TRUE FALSE TRUE
#> 2 FALSE  TRUE TRUE
#> 3  TRUE FALSE TRUE
#> 4  TRUE FALSE TRUE
#> 5 FALSE  TRUE TRUE

by_add_total(NULL, "total", n = 3)

#>   total
#> 1  TRUE
#> 2  TRUE
#> 3  TRUE

attr(by_add_total(
by_hierarchical(factor(c("F", "F", "M", "M")), factor(c("<40", ">40", "<40", ">40"))),
label = "All"
), "header")

#> [[1]]
#>               F                     M               All
#> 
#> [[2]]
#>         <40        >40        <40        >40           
#> 
#> attr(,"nrow")
#> [1] 2
#> attr(,"ncol")
#> [1] 5
#> attr(,"class")
#> [1] "rheader"