Compatability Arg Conventions

Compatability Arg Conventions

compat_args(
  .lst,
  FUN,
  col_by,
  row_by,
  row.name,
  format,
  indent,
  col_wise_args,
  label
)

Arguments

.lst

list. An already-collected list of arguments tot be used instead of the elements of .... Arguments passed via ... will be ignored if this is specified.

FUN

function. Tabulation fucntion. Will be passed subsets of x defined by the combination of col_by and row_by and returns corresponding cell value

col_by

(factor or data.frame if a factor of length nrow(x) that defines which levels in col_by define a column.

row_by

rows in x to take per row in the resulting table

row.name

if NULL then the FUN argument is deparsed and used as row.name of the rrow

format

if FUN does not return a formatted rcell then the format is applied

indent

deprecated.

col_wise_args

a named list containing collections (e.g. vectors or lists) with data elements for each column of the resulting table. The data elements are then passed to the named argument FUN corresponding to the element name of the outer list. Hence, the length and order of each collection must match the levels in col_by. See examples.

label

character(1). A label (not to be confused with the name) for the object/structure.

Value

NULL (this is an argument template dummy function)

See also