Constructor Arg Conventions

Constructor Arg Conventions

constr_args(
  kids,
  cont,
  lev,
  iscontent,
  cinfo,
  labelrow,
  vals,
  cspan,
  label_pos,
  cindent_mod,
  cvar,
  label,
  cextra_args,
  child_names,
  title,
  subtitles,
  main_footer,
  prov_footer,
  footnotes,
  page_title,
  page_prefix
)

Arguments

kids

list. List of direct children.

cont

ElementaryTable. Content table.

lev

integer. Nesting level (roughly, indentation level in practical terms).

iscontent

logical. Is the TableTree/ElementaryTable being constructed the content table for another TableTree.

cinfo

InstantiatedColumnInfo (or NULL). Column structure for the object being created.

labelrow

LabelRow. The LabelRow object to assign to this Table. Consructed from label by default if not specified.

vals

list. cell values for the row

cspan

integer. Column span. 1 indicates no spanning.

label_pos

character(1). Location the variable label should be displayed, Accepts hidden (default for non-analyze row splits), visible, topleft, and - for analyze splits only - default. For analyze calls, default indicates that the variable should be visible if and only if multiple variables are analyzed at the same level of nesting.

cindent_mod

numeric(1). The indent modifier for the content tables generated by this split.

cvar

character(1). The variable, if any, which the content function should accept. Defaults to NA.

label

character(1). A label (not to be confused with the name) for the object/structure.

cextra_args

list. Extra arguments to be passed to the content function when tabulating row group summaries.

child_names

character. Names to be given to the sub splits contained by a compound split (typically a AnalyzeMultiVars split object).

title

character(1). Main title. Ignored for subtables.

subtitles

character. Subtitles. Ignored for subtables.

main_footer

character. Main global (non-referential) footer materials.

prov_footer

character. Provenance-related global footer materials. Generally should not be modified by hand.

footnotes

list or NULL. Referential footnotes to be applied at current level

page_title

character. Page specific title(s).

page_prefix

character(1). Prefix, to be appended with the split value, when forcing pagination between the children of this split/table

Value

NULL (this is an argument template dummy function)

See also