Splits for cutting by values of a numeric variable

Splits for cutting by values of a numeric variable

VarStaticCutSplit(
  var,
  split_label = var,
  cuts,
  cutlabels = NULL,
  cfun = NULL,
  cformat = NULL,
  split_format = NULL,
  split_name = var,
  child_labels = c("default", "visible", "hidden"),
  extra_args = list(),
  indent_mod = 0L,
  cindent_mod = 0L,
  cvar = "",
  cextra_args = list(),
  label_pos = "visible"
)

CumulativeCutSplit(
  var,
  split_label,
  cuts,
  cutlabels = NULL,
  cfun = NULL,
  cformat = NULL,
  split_format = NULL,
  split_name = var,
  child_labels = c("default", "visible", "hidden"),
  extra_args = list(),
  indent_mod = 0L,
  cindent_mod = 0L,
  cvar = "",
  cextra_args = list(),
  label_pos = "visible"
)

VarDynCutSplit(
  var,
  split_label,
  cutfun,
  cutlabelfun = function(x) NULL,
  cfun = NULL,
  cformat = NULL,
  split_format = NULL,
  split_name = var,
  child_labels = c("default", "visible", "hidden"),
  extra_args = list(),
  cumulative = FALSE,
  indent_mod = 0L,
  cindent_mod = 0L,
  cvar = "",
  cextra_args = list(),
  label_pos = "visible"
)

Arguments

var
string, variable name

split_label
string. Label string to be associated with the table generated by the split. Not to be confused with labels assigned to each child (which are based on the data and type of split during tabulation).

cuts
numeric. Cuts to use

cutlabels
character (or NULL). Labels for the cuts

cfun
list/function/NULL. tabulation function(s) for creating content rows. Must accept x or df as first parameter. Must accept labelstr as the second argument. Can optionally accept all optional arguments accepted by analysis functions. See analyze.

cformat
format spec. Format for content rows

split_format
FormatSpec. Default format associated with the split being created.

split_name
string. Name associiated with this split (for pathing, etc)

child_labels
string. One of "default", "visible", "hidden". What should the display behavior be for the labels (ie label rows) of the children of this split. Defaults to "default" which flags the label row as visible only if the child has 0 content rows.

extra_args
list. Extra arguments to be passed to the tabulation function. Element position in thte list corresponds to the children of this split. Named elements in the child-specific lists are ignored if they do not match a formal argument of the ttabulation function.

indent_mod
numeric. Modifier for the default indent position for the structure created by this function(subtable, content table, or row) and all of that structure's children. Defaults to 0, which corresponds to the unmodified default behavior.

cindent_mod
numeric(1). The indent modifier for the content tables generated by this split.

cvar
character(1). The variable, if any, which the content function should accept. Defaults to NA.

cextra_args
list. Extra arguments to be passed to the content function when tabulating row group summaries.

label_pos
character(1). Location the variable label should be displayed, Accepts hidden (default for non-analyze row splits), visible, topleft, and - for analyze splits only - default. For analyze calls, default indicates that the variable should be visible if and only if multiple variables are analyzed at the same level of nesting.

cutfun
function. Function which accepts the full vector of var values and returns cut points to be used (via cut) when splitting data during tabulation

cutlabelfun
function. Function which returns either labels for the cuts or NULL when passed the return value of cutfun

cumulative
logical. Should the cuts be treated as cumulative. Defaults to FALSE

Value

a VarStaticCutSplit, CumulativeCutSplit, or VarDynCutSplit object.