Preprocessing Functions for use in make_split_fun

This function is intended for use as a preprocessing component in make_split_fun, and should not be called directly by end users.

Usage

drop_facet_levels(df, spl, ...)

Arguments

df

data.frame. The incoming data corresponding with the parent facet

spl

Split.

...

dots. This is used internally to pass parameters.

See also