Export as plain text with page break symbol

Export as plain text with page break symbol

export_as_txt(
  tt,
  file = NULL,
  paginate = FALSE,
  ...,
  page_break = "\\s\\n",
  hsep = default_hsep(),
  indent_size = 2
)

Arguments

tt

TableTree (or related class). A TableTree object representing a populated table.

file

character(1). File to write.

paginate

logical(1). Should tt be paginated before writing the file.

...

Passed directly to paginate_table

page_break

character(1). Page break symbol (defaults to outputting "\s").

hsep

character(1). Set of character(s) to be repeated as the separator between the header and body of the table when rendered as text. Defaults to a connected horrizontal line (unicode 2014) in locals that use a UTF charset, and to - elsewhere (with a once per session warning).

indent_size

numeric(1). Number of spaces to use per indent level. Defaults to 2

Value

file (this function is called for the side effect of writing the file.

See also

Examples

lyt <- basic_table() %>%
  split_cols_by("ARM") %>%
  analyze(c("AGE", "BMRKR2", "COUNTRY"))

tbl <- build_table(lyt, ex_adsl)

cat(export_as_txt(tbl, file = NULL, paginate = TRUE, lpp = 8))
#>            A: Drug X   B: Placebo   C: Combination
#> ——————————————————————————————————————————————————
#> AGE                                               
#>   Mean       33.77       35.43          35.43     
#> BMRKR2                                            
#>   LOW         50           45             40      
#>   MEDIUM      37           56             42      
#>   HIGH        47           33             50      
#> \s\n           A: Drug X   B: Placebo   C: Combination
#> ——————————————————————————————————————————————————
#> COUNTRY                                           
#>   CHN         74           81             64      
#>   USA         10           13             17      
#>   BRA         13           7              10      
#>   PAK         12           9              10      
#>   NGA          8           7              11      
#> \s\n           A: Drug X   B: Placebo   C: Combination
#> ——————————————————————————————————————————————————
#> COUNTRY                                           
#>   RUS          5           8              6       
#>   JPN          5           4              9       
#>   GBR          4           3              2       
#>   CAN          3           2              3       
#>   CHE          0           0              0       

if (FALSE) {
tf <- tempfile(fileext = ".txt")
export_as_txt(tbl, file = tf)
system2("cat", tf)
}