Find degenerate (sub)structures within a table

This function returns a list with the row-paths to all structural subtables which contain no data rows (even if they have associated content rows).

Usage

find_degen_struct(tt)

Arguments

tt

(TableTree)
a TableTree object.

Value

A list of character vectors representing the row paths, if any, to degenerate substructures within the table.

See also

Examples

find_degen_struct(rtable("hi"))
#> [[1]]
#> [1] ""
#>