General Argument Conventions

General Argument Conventions

gen_args(
  df,
  alt_counts_df,
  spl,
  pos,
  tt,
  tr,
  verbose,
  colwidths,
  obj,
  x,
  value,
  object,
  path,
  label,
  label_pos,
  cvar,
  topleft,
  page_prefix,
  hsep,
  indent_size,
  ...
)

Arguments

df

dataset (data.frame or tibble)

alt_counts_df

dataset (data.frame or tibble). Alternative full data the rtables framework will use (only) when calculating column counts.

spl

A Split object defining a partitioning or analysis/tabulation of the data.

pos

numeric. Which top-level set of nested splits should the new layout feature be added to. Defaults to the current

tt

TableTree (or related class). A TableTree object representing a populated table.

tr

TableRow (or related class). A TableRow object representing a single row within a populated table.

verbose

logical(1). Should extra debugging messages be shown. Defaults to FALSE.

colwidths

numeric vector. Column widths for use with vertical pagination. Currently ignored.

obj

ANY. The object for the accessor to access or modify

x

An object

value

The new value

object

The object to modify in-place

path

character. A vector path for a position within the structure of a tabletree. Each element represents a subsequent choice amongst the children of the previous choice.

label

character(1). A label (not to be confused with the name) for the object/structure.

label_pos

character(1). Location the variable label should be displayed, Accepts hidden (default for non-analyze row splits), visible, topleft, and - for analyze splits only - default. For analyze calls, default indicates that the variable should be visible if and only if multiple variables are analyzed at the same level of nesting.

cvar

character(1). The variable, if any, which the content function should accept. Defaults to NA.

topleft

character. Override values for the "top left" material to be displayed during printing.

page_prefix

character(1). Prefix, to be appended with the split value, when forcing pagination between the children of this split/table

hsep

character(1). Set of character(s) to be repeated as the separator between the header and body of the table when rendered as text. Defaults to a connected horrizontal line (unicode 2014) in locals that use a UTF charset, and to - elsewhere (with a once per session warning).

indent_size

numeric(1). Number of spaces to use per indent level. Defaults to 2

...

Passed on to methods or tabulation functions.

Value

NULL (this is an argument template dummy function)

See also