Create multiple rows in analysis or summary functions

define the cells that get placed into multiple rows in afun

in_rows(
  ...,
  .list = NULL,
  .names = NULL,
  .labels = NULL,
  .formats = NULL,
  .indent_mods = NULL,
  .cell_footnotes = list(NULL),
  .row_footnotes = list(NULL),
  .aligns = NULL,
  .format_na_strs = NULL
)

Arguments

...

single row defining expressions

.list

list. list cell content, usually rcells, the .list is concatenated to ...

.names

character or NULL. Names of the returned list/structure.

.labels

character or NULL. labels for the defined rows

.formats

character or NULL. Formats for the values

.indent_mods

integer or NULL. Indent modificatons for the defined rows.

.cell_footnotes

list. Referential footnote messages to be associated by name with cells

.row_footnotes

list. Referential footnotes messages to be associated by name with rows

.aligns

character or NULL. Alignments for the cells

.format_na_strs

character or NULL. NA strings for the cells

Value

an RowsVerticalSection object (or NULL). The details of this object should be considered an internal implementation detail.

See also

Examples

in_rows(1, 2, 3, .names = c("a", "b", "c"))
#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>   row_name formatted_cell indent_mod row_label
#> 1        a              1          0         a
#> 2        b              2          0         b
#> 3        c              3          0         c
in_rows(1, 2, 3, .labels = c("a", "b", "c"))
#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>   row_name formatted_cell indent_mod row_label
#> 1        a              1          0         a
#> 2        b              2          0         b
#> 3        c              3          0         c
in_rows(1, 2, 3, .names = c("a", "b", "c"), .labels = c("AAA", "BBB", "CCC"))
#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>   row_name formatted_cell indent_mod row_label
#> 1        a              1          0       AAA
#> 2        b              2          0       BBB
#> 3        c              3          0       CCC

in_rows(.list = list(a = 1, b = 2, c = 3))
#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>   row_name formatted_cell indent_mod row_label
#> 1        a              1          0         a
#> 2        b              2          0         b
#> 3        c              3          0         c
in_rows(1, 2, .list = list(3), .names = c("a", "b", "c"))
#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>   row_name formatted_cell indent_mod row_label
#> 1        a              1          0         a
#> 2        b              2          0         b
#> 3        c              3          0         c

basic_table() %>%
  split_cols_by("ARM") %>%
  analyze("AGE", afun = function(x) {
    in_rows(
       "Mean (sd)" = rcell(c(mean(x), sd(x)), format = "xx.xx (xx.xx)"),
       "Range" = rcell(range(x), format = "xx.xx - xx.xx")
    )
  }) %>%
  build_table(ex_adsl)
#>               A: Drug X      B: Placebo     C: Combination
#> ——————————————————————————————————————————————————————————
#> Mean (sd)   33.77 (6.55)    35.43 (7.90)     35.43 (7.72) 
#> Range       21.00 - 50.00   21.00 - 62.00   20.00 - 69.00