[DEPRECATED] insert rrows at (before) a specific location

This function is deprecated and will be removed in a future release of rtables. Please use insert_row_at_path or label_at_path instead.

Usage

insert_rrow(tbl, rrow, at = 1, ascontent = FALSE)

Arguments

tbl

rtable

rrow

rrow to append to rtable

at

position into which to put the rrow, defaults to beginning (ie 1)

ascontent

logical. Currently ignored.

Value

A TableTree of the same specific class as tbl

Note

Label rows (ie a row with no data values, only a row.name) can only be inserted at positions which do not already contain a label row when there is a non-trivial nested row structure in tbl

Examples

o <- options(warn = 0)
lyt <- basic_table() %>%
    split_cols_by("Species") %>%
    analyze("Sepal.Length")
    
tbl <- build_table(lyt, iris)

insert_rrow(tbl, rrow("Hello World"))
#> Warning: 'insert_rrow' is deprecated.
#> Use 'insert_row_at_path or label_at_path(tab)<-' instead.
#> See help("Deprecated")
#>               setosa   versicolor   virginica
#> —————————————————————————————————————————————
#> Hello World                                  
#> Mean           5.01       5.94        6.59   
insert_rrow(tbl, rrow("Hello World"), at = 2)
#> Warning: 'insert_rrow' is deprecated.
#> Use 'insert_row_at_path or label_at_path(tab)<-' instead.
#> See help("Deprecated")
#>               setosa   versicolor   virginica
#> —————————————————————————————————————————————
#> Mean           5.01       5.94        6.59   
#> Hello World                                  

lyt2 <- basic_table() %>%
    split_cols_by("Species") %>%
    split_rows_by("Species") %>%
    analyze("Sepal.Length")
    
tbl2 <- build_table(lyt2, iris)

insert_rrow(tbl2, rrow("Hello World"))
#> Warning: 'insert_rrow' is deprecated.
#> Use 'insert_row_at_path or label_at_path(tab)<-' instead.
#> See help("Deprecated")
#>               setosa   versicolor   virginica
#> —————————————————————————————————————————————
#> Hello World                                  
#> setosa                                       
#>   Mean         5.01        NA          NA    
#> versicolor                                   
#>   Mean          NA        5.94         NA    
#> virginica                                    
#>   Mean          NA         NA         6.59   
insert_rrow(tbl2, rrow("Hello World"), at = 2)
#> Warning: 'insert_rrow' is deprecated.
#> Use 'insert_row_at_path or label_at_path(tab)<-' instead.
#> See help("Deprecated")
#>                 setosa   versicolor   virginica
#> ———————————————————————————————————————————————
#> setosa                                         
#>   Hello World                                  
#>     Mean         5.01        NA          NA    
#> versicolor                                     
#>   Mean            NA        5.94         NA    
#> virginica                                      
#>   Mean            NA         NA         6.59   
insert_rrow(tbl2, rrow("Hello World"), at = 4)
#> Warning: 'insert_rrow' is deprecated.
#> Use 'insert_row_at_path or label_at_path(tab)<-' instead.
#> See help("Deprecated")
#>                 setosa   versicolor   virginica
#> ———————————————————————————————————————————————
#> setosa                                         
#>   Mean           5.01        NA          NA    
#> versicolor                                     
#>   Hello World                                  
#>     Mean          NA        5.94         NA    
#> virginica                                      
#>   Mean            NA         NA         6.59   

insert_rrow(tbl2, rrow("new row", 5, 6, 7))
#> Warning: 'insert_rrow' is deprecated.
#> Use 'insert_row_at_path or label_at_path(tab)<-' instead.
#> See help("Deprecated")
#>              setosa   versicolor   virginica
#> ————————————————————————————————————————————
#> new row        5          6            7    
#> setosa                                      
#>   Mean        5.01        NA          NA    
#> versicolor                                  
#>   Mean         NA        5.94         NA    
#> virginica                                   
#>   Mean         NA         NA         6.59   

insert_rrow(tbl2, rrow("new row", 5, 6, 7), at = 3)
#> Warning: 'insert_rrow' is deprecated.
#> Use 'insert_row_at_path or label_at_path(tab)<-' instead.
#> See help("Deprecated")
#>              setosa   versicolor   virginica
#> ————————————————————————————————————————————
#> setosa                                      
#>   Mean        5.01        NA          NA    
#>   new row      5          6            7    
#> versicolor                                  
#>   Mean         NA        5.94         NA    
#> virginica                                   
#>   Mean         NA         NA         6.59   

options(o)