Layouting Function Arg Conventions

Layouting Function Arg Conventions

lyt_args(
  lyt,
  var,
  vars,
  label,
  labels_var,
  varlabels,
  varnames,
  split_format,
  nested,
  format,
  cfun,
  cformat,
  split_fun,
  split_name,
  split_label,
  afun,
  inclNAs,
  valorder,
  ref_group,
  compfun,
  label_fstr,
  child_labels,
  extra_args,
  name,
  cuts,
  cutlabels,
  cutfun,
  cutlabelfun,
  cumulative,
  incl_all,
  indent_mod,
  show_labels,
  visible_label,
  var_labels,
  cvar,
  table_names,
  topleft
)

Arguments

lyt
layout object pre-data used for tabulation

var
string, variable name

vars
character vector. Multiple variable names.

label
character(1). A label (not to be confused with the name) for the object/structure.

labels_var
string, name of variable containing labels to be displayed for the values of var

varlabels
character vector. Labels for vars

varnames
character vector. Names for vars which will appear in pathing. When vars are all unique this will be the variable names. If not, these will be variable names with suffixes as necessary to enforce uniqueness.

split_format
format spec. Format associated with this split.

nested
boolean, Add this as a new top-level split (defining a new subtable directly under root). Defaults to FALSE

format
FormatSpec. Format associated with this split. Formats can be declared via strings ("xx.x") or function. In cases such as analyze calls, they can character vectors or lists of functions.

cfun
list/function/NULL. tabulation function(s) for creating content rows. Must accept x or df as first parameter. Must accept labelstr as the second argument. Can optionally accept all optional arguments accepted by analysis functions. See analyze.

cformat
format spec. Format for content rows

split_fun
function/NULL. custom splitting function

split_name
string. Name associiated with this split (for pathing, etc)

split_label
string. Label string to be associated with the table generated by the split. Not to be confused with labels assigned to each child (which are based on the data and type of split during tabulation).

afun
function. Analysis function, must take x or df as its first parameter. Can optionally take other parameters which will be populated by the tabulation framework. See Details in analyze.

inclNAs
boolean. Should observations with NA in the var variable(s) be included when performing this analysis. Defaults to FALSE

valorder
character vector. Order that the split children should appear in resulting table.

ref_group
character. Value of var to be taken as the ref_group/control to be compared against.

compfun
function/string. The comparison function which accepts the analysis function outputs for two different partitions and returns a single value. Defaults to subraction. If a string, taken as the name of a function.

label_fstr
string. An sprintf style format string containing. For non-comparison splits, it can contain up to one "%s" which takes the current split value and generates the row/column label. Comparison-based splits it can contain up to two "%s".

child_labels
string. One of "default", "visible", "hidden". What should the display behavior be for the labels (ie label rows) of the children of this split. Defaults to "default" which flags the label row as visible only if the child has 0 content rows.

extra_args
list. Extra arguments to be passed to the tabulation function. Element position in thte list corresponds to the children of this split. Named elements in the child-specific lists are ignored if they do not match a formal argument of the ttabulation function.

name
character(1). Name of the split/table/row being creatted. Defaults to same as the corresponding llabel, but is not required to be.

cuts
numeric. Cuts to use

cutlabels
character (or NULL). Labels for the cuts

cutfun
function. Function which accepts the full vector of var values and returns cut points to be used (via cut) when splitting data during tabulation

cutlabelfun
function. Function which returns either labels for the cuts or NULL when passed the return value of cutfun

cumulative
logical. Should the cuts be treated as cumulative. Defaults to FALSE

incl_all
logical(1). Should an "all" comparison column be created. Defaults to FALSE.

indent_mod
numeric. Modifier for the default indent position for the structure created by this function(subtable, content table, or row) and all of that structure's children. Defaults to 0, which corresponds to the unmodified default behavior.

show_labels
character(1). Should the variable labels for corresponding to the variable(s) in vars be visible in the resulting table.

visible_label
logical(1). Should the split label be visible in the resulting table.

var_labels
character. Variable labels for 1 or more variables

cvar
character(1). The variable, if any, which the content function should accept. Defaults to NA.

table_names
character. Names for the tables representing each atomic analysis. Defaults to var.

topleft
character. Override values for the "top left" material to be displayed during printing.

See also