Manual column declaration

Manual column declaration

Usage

manual_cols(..., .lst = list(...))

Arguments

...

One or more vectors of levels to appear in the column space. If more than one set of levels is given, the values of the second are nested within each value of the first, and so on.

.lst

A list of sets of levels, by default populated via list(...).

Value

An InstantiatedColumnInfo object, suitable for use declaring the column structure for a manually constructed table.

Author

Gabriel Becker

Examples

# simple one level column space
rows <- lapply(1:5, function(i) {
   DataRow(rep(i, times  = 3))})
tbl <- TableTree(kids = rows, cinfo = manual_cols(split = c("a", "b", "c")))
tbl
#>    a   b   c
#> ————————————
#>    1   1   1
#>    2   2   2
#>    3   3   3
#>    4   4   4
#>    5   5   5

# manually declared nesting
tbl2 <- TableTree(kids = list(DataRow(as.list(1:4))),
                 cinfo = manual_cols(Arm = c("Arm A", "Arm B"),
                                     Gender = c("M", "F")))
tbl2
#>     Arm A      Arm B  
#>     M    F     M    F 
#> ——————————————————————
#>     1    2     3    4