Transform rtable to a list of matrices which can be used for outputting

Although rtables are represented as a tree data structure when outputting the table to ASCII or HTML it is useful to map the rtable to an in between state with the formatted cells in a matrix form.

matrix_form(tt, indent_rownames = FALSE)

Arguments

tt
TableTree (or related class). A TableTree object representing a populated table.

indent_rownames
logical(1), if TRUE the column with the row names in the strings matrix of has indented row names (strings pre-fixed)

Examples

library(dplyr)

iris2 <- iris %>%
  group_by(Species) %>%
  mutate(group = as.factor(rep_len(c("a", "b"), length.out = n()))) %>%
  ungroup()

l <- basic_table() %>%
  split_cols_by("Species") %>%
  split_cols_by("group") %>%
  analyze(c("Sepal.Length", "Petal.Width"), afun = list_wrap_x(summary) , format = "xx.xx")

l

#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> Species (lvls) -> group (lvls) 
#> 
#> Row-Split Structure:
#> ( (** multivar analysis **) -> Sepal.Length, Petal.Width (** multivar analysis **) -> ) (** multivar analysis **) 
#> 

tbl <- build_table(l, iris2)

matrix_form(tbl)

#> $strings
#>       [,1]           [,2]     [,3]     [,4]         [,5]         [,6]       
#>  [1,] ""             "setosa" "setosa" "versicolor" "versicolor" "virginica"
#>  [2,] ""             "a"      "b"      "a"          "b"          "a"        
#>  [3,] "Sepal.Length" ""       ""       ""           ""           ""         
#>  [4,] "Min."         "4.4"    "4.3"    "5"          "4.9"        "4.9"      
#>  [5,] "1st Qu."      "4.8"    "4.8"    "5.6"        "5.6"        "6.2"      
#>  [6,] "Median"       "5"      "5"      "5.9"        "5.9"        "6.5"      
#>  [7,] "Mean"         "5.02"   "4.99"   "5.99"       "5.88"       "6.5"      
#>  [8,] "3rd Qu."      "5.3"    "5.1"    "6.4"        "6.1"        "6.7"      
#>  [9,] "Max."         "5.8"    "5.7"    "7"          "6.7"        "7.7"      
#> [10,] "Petal.Width"  ""       ""       ""           ""           ""         
#> [11,] "Min."         "0.1"    "0.1"    "1"          "1"          "1.4"      
#> [12,] "1st Qu."      "0.2"    "0.2"    "1.2"        "1.2"        "1.9"      
#> [13,] "Median"       "0.2"    "0.2"    "1.3"        "1.3"        "2.1"      
#> [14,] "Mean"         "0.23"   "0.26"   "1.35"       "1.3"        "2.08"     
#> [15,] "3rd Qu."      "0.2"    "0.3"    "1.5"        "1.4"        "2.3"      
#> [16,] "Max."         "0.4"    "0.6"    "1.8"        "1.7"        "2.5"      
#>       [,7]       
#>  [1,] "virginica"
#>  [2,] "b"        
#>  [3,] ""         
#>  [4,] "5.6"      
#>  [5,] "6.3"      
#>  [6,] "6.5"      
#>  [7,] "6.67"     
#>  [8,] "7.2"      
#>  [9,] "7.9"      
#> [10,] ""         
#> [11,] "1.5"      
#> [12,] "1.8"      
#> [13,] "2"        
#> [14,] "1.98"     
#> [15,] "2.2"      
#> [16,] "2.5"      
#> 
#> $spans
#>         setosa.a setosa.b versicolor.a versicolor.b virginica.a virginica.b
#>  [1,] 1        2        2            2            2           2           2
#>  [2,] 1        1        1            1            1           1           1
#>  [3,] 1        1        1            1            1           1           1
#>  [4,] 1        1        1            1            1           1           1
#>  [5,] 1        1        1            1            1           1           1
#>  [6,] 1        1        1            1            1           1           1
#>  [7,] 1        1        1            1            1           1           1
#>  [8,] 1        1        1            1            1           1           1
#>  [9,] 1        1        1            1            1           1           1
#> [10,] 1        1        1            1            1           1           1
#> [11,] 1        1        1            1            1           1           1
#> [12,] 1        1        1            1            1           1           1
#> [13,] 1        1        1            1            1           1           1
#> [14,] 1        1        1            1            1           1           1
#> [15,] 1        1        1            1            1           1           1
#> [16,] 1        1        1            1            1           1           1
#> 
#> $aligns
#>       [,1]   [,2]     [,3]     [,4]     [,5]     [,6]     [,7]    
#>  [1,] "left" "center" "center" "center" "center" "center" "center"
#>  [2,] "left" "center" "center" "center" "center" "center" "center"
#>  [3,] "left" "center" "center" "center" "center" "center" "center"
#>  [4,] "left" "center" "center" "center" "center" "center" "center"
#>  [5,] "left" "center" "center" "center" "center" "center" "center"
#>  [6,] "left" "center" "center" "center" "center" "center" "center"
#>  [7,] "left" "center" "center" "center" "center" "center" "center"
#>  [8,] "left" "center" "center" "center" "center" "center" "center"
#>  [9,] "left" "center" "center" "center" "center" "center" "center"
#> [10,] "left" "center" "center" "center" "center" "center" "center"
#> [11,] "left" "center" "center" "center" "center" "center" "center"
#> [12,] "left" "center" "center" "center" "center" "center" "center"
#> [13,] "left" "center" "center" "center" "center" "center" "center"
#> [14,] "left" "center" "center" "center" "center" "center" "center"
#> [15,] "left" "center" "center" "center" "center" "center" "center"
#> [16,] "left" "center" "center" "center" "center" "center" "center"
#> 
#> $display
#>            setosa.a setosa.b versicolor.a versicolor.b virginica.a virginica.b
#>  [1,] TRUE     TRUE    FALSE         TRUE        FALSE        TRUE       FALSE
#>  [2,] TRUE     TRUE     TRUE         TRUE         TRUE        TRUE        TRUE
#>  [3,] TRUE     TRUE     TRUE         TRUE         TRUE        TRUE        TRUE
#>  [4,] TRUE     TRUE     TRUE         TRUE         TRUE        TRUE        TRUE
#>  [5,] TRUE     TRUE     TRUE         TRUE         TRUE        TRUE        TRUE
#>  [6,] TRUE     TRUE     TRUE         TRUE         TRUE        TRUE        TRUE
#>  [7,] TRUE     TRUE     TRUE         TRUE         TRUE        TRUE        TRUE
#>  [8,] TRUE     TRUE     TRUE         TRUE         TRUE        TRUE        TRUE
#>  [9,] TRUE     TRUE     TRUE         TRUE         TRUE        TRUE        TRUE
#> [10,] TRUE     TRUE     TRUE         TRUE         TRUE        TRUE        TRUE
#> [11,] TRUE     TRUE     TRUE         TRUE         TRUE        TRUE        TRUE
#> [12,] TRUE     TRUE     TRUE         TRUE         TRUE        TRUE        TRUE
#> [13,] TRUE     TRUE     TRUE         TRUE         TRUE        TRUE        TRUE
#> [14,] TRUE     TRUE     TRUE         TRUE         TRUE        TRUE        TRUE
#> [15,] TRUE     TRUE     TRUE         TRUE         TRUE        TRUE        TRUE
#> [16,] TRUE     TRUE     TRUE         TRUE         TRUE        TRUE        TRUE
#> 
#> $row_info
#>                              name        label indent depth  rowtype indent_mod
#> Sepal.Length1        Sepal.Length Sepal.Length      0     1 LabelRow          0
#> Sepal.Length.Min.            Min.         Min.      1     2 TableRow          0
#> Sepal.Length.1st Qu.      1st Qu.      1st Qu.      1     2 TableRow          0
#> Sepal.Length.Median        Median       Median      1     2 TableRow          0
#> Sepal.Length.Mean            Mean         Mean      1     2 TableRow          0
#> Sepal.Length.3rd Qu.      3rd Qu.      3rd Qu.      1     2 TableRow          0
#> Sepal.Length.Max.            Max.         Max.      1     2 TableRow          0
#> Petal.Width1          Petal.Width  Petal.Width      0     1 LabelRow          0
#> Petal.Width.Min.             Min.         Min.      1     2 TableRow          0
#> Petal.Width.1st Qu.       1st Qu.      1st Qu.      1     2 TableRow          0
#> Petal.Width.Median         Median       Median      1     2 TableRow          0
#> Petal.Width.Mean             Mean         Mean      1     2 TableRow          0
#> Petal.Width.3rd Qu.       3rd Qu.      3rd Qu.      1     2 TableRow          0
#> Petal.Width.Max.             Max.         Max.      1     2 TableRow          0
#>                      level
#> Sepal.Length1            1
#> Sepal.Length.Min.        2
#> Sepal.Length.1st Qu.     2
#> Sepal.Length.Median      2
#> Sepal.Length.Mean        2
#> Sepal.Length.3rd Qu.     2
#> Sepal.Length.Max.        2
#> Petal.Width1             1
#> Petal.Width.Min.         2
#> Petal.Width.1st Qu.      2
#> Petal.Width.Median       2
#> Petal.Width.Mean         2
#> Petal.Width.3rd Qu.      2
#> Petal.Width.Max.         2
#> 
#> attr(,"nrow_header")
#> [1] 2