Pagination of a TableTree

Pagination of a TableTree

pag_tt_indices(
  tt,
  lpp = 15,
  min_siblings = 2,
  nosplitin = character(),
  colwidths = NULL,
  verbose = FALSE
)

paginate_table(
  tt,
  lpp = 15,
  min_siblings = 2,
  nosplitin = character(),
  colwidths = NULL,
  verbose = FALSE
)

Arguments

tt
TableTree (or related class). A TableTree object representing a populated table.

lpp
numeric. Maximum lines per page including (re)printed header and context rows

min_siblings
numeric. Minimum sibling rows which must appear on either side of pagination row for a mid-subtable split to be valid. Defaults to 2.

nosplitin
character. List of names of sub-tables where page-breaks are not allowed, regardless of other considerations. Defaults to none.

colwidths
numeric vector. Column widths for use with vertical pagination. Currently ignored.

verbose
logical(1). Should extra debugging messages be shown. Defaults to FALSE.

Value

for pag_tt_indices a list of paginated-groups of row-indices of tt. For paginate_table, The subtables defined by subsetting by the indices defined by pag_tt_indices.

Note

This is our first take on pagination. We will refine pagination in subsequent releases. Currently only pagination in the row space work. Pagination in the column space will be added in the future.

Examples

s_summary <- function(x) {
 if (is.numeric(x)) {
     in_rows(
         "n" = rcell(sum(!is.na(x)), format = "xx"),
         "Mean (sd)" = rcell(c(mean(x, na.rm = TRUE), sd(x, na.rm = TRUE)),
                             format = "xx.xx (xx.xx)"),
         "IQR" = rcell(IQR(x, na.rm = TRUE), format = "xx.xx"),
         "min - max" = rcell(range(x, na.rm = TRUE), format = "xx.xx - xx.xx")
     )
 } else if (is.factor(x)) {

     vs <- as.list(table(x))
     do.call(in_rows, lapply(vs, rcell, format = "xx"))

 } else (
     stop("type not supported")
 )
}

lyt <- basic_table() %>%
split_cols_by(var = "ARM") %>%
    analyze(c("AGE", "SEX", "BEP01FL", "BMRKR1", "BMRKR2", "COUNTRY"), afun = s_summary)

tbl <- build_table(lyt, ex_adsl)
tbl

#>                       A: Drug X      B: Placebo    C: Combination
#> -----------------------------------------------------------------
#> AGE                                                              
#>   n                      134            134             132      
#>   Mean (sd)          33.77 (6.55)   35.43 (7.9)     35.43 (7.72) 
#>   IQR                     11             10              10      
#>   min - max            21 - 50        21 - 62         20 - 69    
#> SEX                                                              
#>   F                       79             77              66      
#>   M                       51             55              60      
#>   U                       3              2               4       
#>   UNDIFFERENTIATED        1              0               2       
#> BEP01FL                                                          
#>   Y                       68             63              66      
#>   N                       66             71              66      
#> BMRKR1                                                           
#>   n                      134            134             132      
#>   Mean (sd)          5.97 (3.55)     5.7 (3.31)     5.62 (3.49)  
#>   IQR                    4.16           4.06            3.88     
#>   min - max          0.41 - 17.67   0.65 - 14.24    0.17 - 21.39 
#> BMRKR2                                                           
#>   LOW                     50             45              40      
#>   MEDIUM                  37             56              42      
#>   HIGH                    47             33              50      
#> COUNTRY                                                          
#>   CHN                     74             81              64      
#>   USA                     10             13              17      
#>   BRA                     13             7               10      
#>   PAK                     12             9               10      
#>   NGA                     8              7               11      
#>   RUS                     5              8               6       
#>   JPN                     5              4               9       
#>   GBR                     4              3               2       
#>   CAN                     3              2               3       
#>   CHE                     0              0               0       

nrow(tbl)

#> [1] 33

row_paths_summary(tbl)

#> rowname               node_class    path                 
#> ---------------------------------------------------------
#> AGE                   LabelRow      AGE                  
#>   n                   DataRow       AGE, n               
#>   Mean (sd)           DataRow       AGE, Mean (sd)       
#>   IQR                 DataRow       AGE, IQR             
#>   min - max           DataRow       AGE, min - max       
#> SEX                   LabelRow      SEX                  
#>   F                   DataRow       SEX, F               
#>   M                   DataRow       SEX, M               
#>   U                   DataRow       SEX, U               
#>   UNDIFFERENTIATED    DataRow       SEX, UNDIFFERENTIATED
#> BEP01FL               LabelRow      BEP01FL              
#>   Y                   DataRow       BEP01FL, Y           
#>   N                   DataRow       BEP01FL, N           
#> BMRKR1                LabelRow      BMRKR1               
#>   n                   DataRow       BMRKR1, n            
#>   Mean (sd)           DataRow       BMRKR1, Mean (sd)    
#>   IQR                 DataRow       BMRKR1, IQR          
#>   min - max           DataRow       BMRKR1, min - max    
#> BMRKR2                LabelRow      BMRKR2               
#>   LOW                 DataRow       BMRKR2, LOW          
#>   MEDIUM              DataRow       BMRKR2, MEDIUM       
#>   HIGH                DataRow       BMRKR2, HIGH         
#> COUNTRY               LabelRow      COUNTRY              
#>   CHN                 DataRow       COUNTRY, CHN         
#>   USA                 DataRow       COUNTRY, USA         
#>   BRA                 DataRow       COUNTRY, BRA         
#>   PAK                 DataRow       COUNTRY, PAK         
#>   NGA                 DataRow       COUNTRY, NGA         
#>   RUS                 DataRow       COUNTRY, RUS         
#>   JPN                 DataRow       COUNTRY, JPN         
#>   GBR                 DataRow       COUNTRY, GBR         
#>   CAN                 DataRow       COUNTRY, CAN         
#>   CHE                 DataRow       COUNTRY, CHE         

tbls <- paginate_table(tbl)

w_tbls <- propose_column_widths(tbl) # so that we have the same column widths

tmp <- lapply(tbls, print, widths = w_tbls)

#>                       A: Drug X      B: Placebo    C: Combination
#> -----------------------------------------------------------------
#> AGE                                                              
#>   n                      134            134             132      
#>   Mean (sd)          33.77 (6.55)   35.43 (7.9)     35.43 (7.72) 
#>   IQR                     11             10              10      
#>   min - max            21 - 50        21 - 62         20 - 69    
#> SEX                                                              
#>   F                       79             77              66      
#>   M                       51             55              60      
#>   U                       3              2               4       
#>   UNDIFFERENTIATED        1              0               2       
#> BEP01FL                                                          
#>   Y                       68             63              66      
#>   N                       66             71              66      
#>                       A: Drug X      B: Placebo    C: Combination
#> -----------------------------------------------------------------
#> BMRKR1                                                           
#>   n                      134            134             132      
#>   Mean (sd)          5.97 (3.55)     5.7 (3.31)     5.62 (3.49)  
#>   IQR                    4.16           4.06            3.88     
#>   min - max          0.41 - 17.67   0.65 - 14.24    0.17 - 21.39 
#> BMRKR2                                                           
#>   LOW                     50             45              40      
#>   MEDIUM                  37             56              42      
#>   HIGH                    47             33              50      
#> COUNTRY                                                          
#>   CHN                     74             81              64      
#>   USA                     10             13              17      
#>   BRA                     13             7               10      
#>                       A: Drug X      B: Placebo    C: Combination
#> -----------------------------------------------------------------
#> COUNTRY                                                          
#>   PAK                     12             9               10      
#>   NGA                     8              7               11      
#>   RUS                     5              8               6       
#>   JPN                     5              4               9       
#>   GBR                     4              3               2       
#>   CAN                     3              2               3       
#>   CHE                     0              0               0       

tmp <- lapply(tbls, function(tbli) {
  cat(toString(tbli, widths = w_tbls))
  cat("\n\n")
  cat("~~~~ PAGE BREAK ~~~~")
  cat("\n\n")
})

#>                       A: Drug X      B: Placebo    C: Combination
#> -----------------------------------------------------------------
#> AGE                                                              
#>   n                      134            134             132      
#>   Mean (sd)          33.77 (6.55)   35.43 (7.9)     35.43 (7.72) 
#>   IQR                     11             10              10      
#>   min - max            21 - 50        21 - 62         20 - 69    
#> SEX                                                              
#>   F                       79             77              66      
#>   M                       51             55              60      
#>   U                       3              2               4       
#>   UNDIFFERENTIATED        1              0               2       
#> BEP01FL                                                          
#>   Y                       68             63              66      
#>   N                       66             71              66      
#> 
#> 
#> ~~~~ PAGE BREAK ~~~~
#> 
#>                       A: Drug X      B: Placebo    C: Combination
#> -----------------------------------------------------------------
#> BMRKR1                                                           
#>   n                      134            134             132      
#>   Mean (sd)          5.97 (3.55)     5.7 (3.31)     5.62 (3.49)  
#>   IQR                    4.16           4.06            3.88     
#>   min - max          0.41 - 17.67   0.65 - 14.24    0.17 - 21.39 
#> BMRKR2                                                           
#>   LOW                     50             45              40      
#>   MEDIUM                  37             56              42      
#>   HIGH                    47             33              50      
#> COUNTRY                                                          
#>   CHN                     74             81              64      
#>   USA                     10             13              17      
#>   BRA                     13             7               10      
#> 
#> 
#> ~~~~ PAGE BREAK ~~~~
#> 
#>                       A: Drug X      B: Placebo    C: Combination
#> -----------------------------------------------------------------
#> COUNTRY                                                          
#>   PAK                     12             9               10      
#>   NGA                     8              7               11      
#>   RUS                     5              8               6       
#>   JPN                     5              4               9       
#>   GBR                     4              3               2       
#>   CAN                     3              2               3       
#>   CHE                     0              0               0       
#> 
#> 
#> ~~~~ PAGE BREAK ~~~~
#>