Transform TableTree object to Path-Enriched data.frame

Transform TableTree object to Path-Enriched data.frame

path_enriched_df(tt, path_fun = collapse_path, value_fun = collapse_values)

Arguments

tt
TableTree (or related class). A TableTree object representing a populated table.

path_fun
function. Function to transform paths into single-string row/column names.

value_fun
function. Functiont to transform cell values into cells of the data.frame. Defaults to collapse_values which creates strings where multi-valued cells are collapsed together, separated by |.

Value

A data frame of tt's cell values (processed by value_fun, with columns named by the full column paths (processed by path_fun and an additional row_path column with the row paths (processed by by path_fun).

Examples

lyt <- basic_table() %>%
  split_cols_by("ARM") %>%
  analyze(c("AGE", "BMRKR2"))

tbl <- build_table(lyt, ex_adsl)
path_enriched_df(tbl)
#>                      row_path ARM|A: Drug X ARM|B: Placebo ARM|C: Combination
#> 1      ma_AGE_BMRKR2|AGE|Mean      33.76866             50                 37
#> 2    ma_AGE_BMRKR2|BMRKR2|LOW      33.76866             50                 37
#> 3 ma_AGE_BMRKR2|BMRKR2|MEDIUM      33.76866             50                 37
#> 4   ma_AGE_BMRKR2|BMRKR2|HIGH      33.76866             50                 37
#>   NA       NA NA NA NA       NA NA NA NA
#> 1 47 35.43284 45 56 33 35.43182 40 42 50
#> 2 47 35.43284 45 56 33 35.43182 40 42 50
#> 3 47 35.43284 45 56 33 35.43182 40 42 50
#> 4 47 35.43284 45 56 33 35.43182 40 42 50