Cell value constructors

Construct a cell value and associate formatting, labeling, indenting, and column spanning information with it.

rcell(
  x,
  format = NULL,
  colspan = 1L,
  label = NULL,
  indent_mod = NULL,
  footnotes = NULL,
  align = NULL,
  format_na_str = NULL
)

non_ref_rcell(
  x,
  is_ref,
  format = NULL,
  colspan = 1L,
  label = NULL,
  indent_mod = NULL,
  refval = NULL,
  align = "center",
  format_na_str = NULL
)

Arguments

x

ANY. Cell value

format

if FUN does not return a formatted rcell then the format is applied

colspan

integer(1). Columnspan value.

label

character(1). Label or Null. If non-null, it will be looked at when determining row labels.

indent_mod

numeric. Modifier for the default indent position for the structure created by this function(subtable, content table, or row) and all of that structure's children. Defaults to 0, which corresponds to the unmodified default behavior.

footnotes

list or NULL. Referential footnote messages for the cell.

align

character(1). Alignment the value should be rendered with. See rtables_aligns for currently supported alignments.

format_na_str

character(1). String which should be displayed when formatted if this cell's value(s) are all NA.

is_ref

logical(1). Are we in the reference column (ie .in_ref_col shoul be passed to this argument)

refval

ANY. Value to use when in the reference column. Defaults to NULL

Value

An object representing the value within a single cell within a populated table. The underlying structure of this object is an implementation detail and sholud not be relied upon beyond calling accessors for the class.

Details

non_ref_rcell provides the common blank for cells in the reference column, this value otherwise, and should be passed the value of .in_ref_col when it is used.

Note

currently column spanning is only supported for defining header structure.