Create a header

Create a header

Usage

rheader(..., format = "xx", .lst = NULL)

Arguments

...

row specifications (either as character vectors or the output from rrow or DataRow, LabelRow, etc.

format

character(1) or function. The format label (string) or formatter function to apply to the cell values passed via .... See list_valid_format_labels for currently supported format labels.

.lst

list. An already-collected list of arguments to be used instead of the elements of .... Arguments passed via ... will be ignored if this is specified.

Value

a InstantiatedColumnInfo object.

See also

Examples

h1 <- rheader(c("A", "B", "C"))

h2 <- rheader(
  rrow(NULL, rcell("group 1", colspan = 2), rcell("group 2", colspan = 2)),
  rrow(NULL, "A", "B", "A", "B")
)

h1
#> An InstantiatedColumnInfo object
#> Columns:
#> A (mnl)
#> B (mnl)
#> C (mnl)
#> 

h2
#> An InstantiatedColumnInfo object
#> Columns:
#> group 1 (mnl) -> A (mnl)
#> group 1 (mnl) -> B (mnl)
#> group 2 (mnl) -> A (mnl)
#> group 2 (mnl) -> B (mnl)
#>