Row classes and constructors

Row classes and constructors

Row constructors and Classes

LabelRow(
  lev = 1L,
  label = "",
  name = label,
  vis = !is.na(label) && nzchar(label),
  cinfo = EmptyColInfo,
  indent_mod = 0L
)

.tablerow(
  vals = list(),
  name = "",
  lev = 1L,
  label = name,
  cspan = rep(1L, length(vals)),
  cinfo = EmptyColInfo,
  var = NA_character_,
  format = NULL,
  klass,
  indent_mod = 0L
)

DataRow(...)

ContentRow(...)

Arguments

lev
integer. Nesting level (roughly, indentation level in practical terms).

label
character(1). A label (not to be confused with the name) for the object/structure.

name
character(1). Name of the split/table/row being creatted. Defaults to same as the corresponding llabel, but is not required to be.

vis
logical. Should the row be visible (LabelRow only).

cinfo
InstantiatedColumnInfo (or NULL). Column structure for the object being created.

indent_mod
numeric. Modifier for the default indent position for the structure created by this function(subtable, content table, or row) and all of that structure's children. Defaults to 0, which corresponds to the unmodified default behavior.

vals
list. cell values for the row

cspan
integer. Column span. 1 indicates no spanning.

var
string, variable name

format
FormatSpec. Format associated with this split. Formats can be declared via strings ("xx.x") or function. In cases such as analyze calls, they can character vectors or lists of functions.

klass
Internal detail.

...
passed to shared constructor (.tablerow).

Author

Gabriel Becker