Create an rtable row

Create an rtable row

Usage

rrow(row.name = "", ..., format = NULL, indent = 0, inset = 0L)

Arguments

row.name

(string or NULL)
row name. If NULL, an empty string is used as row.name of the rrow().

...

cell values.

format

(string, function, or list)
the format label (string) or formatter function to apply to the cell values passed via .... See formatters::list_valid_format_labels() for currently supported format labels.

indent

inset

(integer(1))
the table inset for the row or table being constructed. See formatters::table_inset() for details.

Value

A row object of the context-appropriate type (label or data).

See also

Examples

rrow("ABC", c(1, 2), c(3, 2), format = "xx (xx.%)")
#> [DataRow indent_mod 0]: ABC   1 (200%)   3 (200%)
rrow("")
#> [LabelRow indent_mod 0]: