Split data.frame and apply functions

Split data.frame and apply functions

# S3 method for data.frame
rtabulate(x, row_by_var = no_by("row_1"),
  col_by_var = no_by("col_1"), FUN = nrow, ..., row_col_data_args = FALSE,
  format = "xx", indent = 0)

Arguments

x
a vecor

row_by_var
name of factor variable in x

col_by_var
name of factor variable in x

FUN
a function that processes the cell data, if row_data_arg is set to TRUE the a second argument with the row data is passed to FUN

...
arguments passed to FUN

row_col_data_args
boolean, if TRUE then FUN is called with the first three arguments being the cell, row, and column data, respectively

format
if FUN does not return a formatted rcell then the format is applied

indent
non-negative integer where 0 means that the row should not be indented

Value

an rtable project

Examples

df <- expand.grid(aaa = factor(c("A", "B")), bbb = factor(c("X", "Y", "Z")))
df <- rbind(df, df)
df$val <- 1:nrow(df)

rtabulate(
  x = df,
  row_by_var = "aaa",
  col_by_var = "bbb",
  FUN = function(x) {
     sum(x$val)
  }
)

#>          X         Y         Z 
#> -------------------------------
#> A        8         12        16
#> B        10        14        18

rtabulate(
  x = iris,
  row_by_var = no_by("sum"),
  col_by_var = "Species",
  FUN = function(x) sum(x$Sepal.Length)
)

#>              setosa          versicolor        virginica 
#> ---------------------------------------------------------
#> sum          250.3             296.8             329.4   

rtabulate(
  x = iris,
  row_by_var = "Species",
  col_by_var = no_by("sum"),
  FUN = function(x) sum(x$Sepal.Length)
)

#>                    sum 
#> -----------------------
#> setosa            250.3
#> versicolor        296.8
#> virginica         329.4

tbl <- rtabulate(
  x = iris,
  FUN = function(cell_data) c(sum(cell_data$Sepal.Length), sd(cell_data$Sepal.Length)),
  format = "xx.xx (xx.xx%)"
)

tbl

#>                  col_1     
#> ---------------------------
#> row_1        876.5 (82.81%)

row.names(tbl)

#>   row_1 
#> "row_1" 

row.names(tbl) <- "Sum of Sepal Length"

tbl

#>                                col_1     
#> -----------------------------------------
#> Sum of Sepal Length        876.5 (82.81%)

iris2 <- iris
iris2$fsl5 <- factor(iris$Sepal.Length > 5, levels = c(TRUE, FALSE),
    labels = c("S.L > 5", "S.L <= 5"))

tbl <- rtabulate(
  x = iris2,
  row_by_var = "fsl5",
  col_by_var = "Species",
  FUN = function(x_cell, x_row, x_col) {
    if (nrow(x_cell) < 10) {
      rcell("-")
    } else {
      fit <- lm(Sepal.Length ~ Petal.Width, data = x_cell)
      m_col <- mean(x_col$Sepal.Length)
      m_row <- mean(x_row$Sepal.Length)

      rcell(list(fit, m_col, m_row), format = function(x, output) {
        paste("df:", x[[1]]$df.residual,", and", round(x[[2]],1), ", and", round(x[[3]],2))
      })
    }
  } ,
  row_col_data_args = TRUE
)
tbl

#>                           setosa                           versicolor                          virginica         
#> -----------------------------------------------------------------------------------------------------------------
#> S.L > 5          df: 20 , and 5 , and 6.13         df: 45 , and 5.9 , and 6.13        df: 47 , and 6.6 , and 6.13
#> S.L <= 5         df: 26 , and 5 , and 4.79                      -                                  -