Tabulate Factors

By default each cell reports the number of observations in each level of x.

# S3 method for factor
rtabulate(x, col_by = by_all("All"), FUN = length,
  ..., useNA = c("ifany", "no", "always"), format = NULL, indent = 0,
  col_wise_args = NULL)

Arguments

x
a vector

col_by
(factor or data.frame if a factor of length nrow(x) that defines which levels in col_by define a column. can use by_factor_to_matrix to create a matrix from a factor to use non-disjoint columns can use by_all to have a column with all rows, alternatively look at by_add_total)

FUN
a function that processes the cell data

...
arguments passed to FUN

useNA
either one of ("no", "ifany", "always"). If "no" then NA values in x get dropped. When "ifany" is used a row for NA values is included in the summary if any NAs exist in x. For option "always" NA values are always included in the summary even if none exist in x.

format
if FUN does not return a formatted rcell then the format is applied

indent
non-negative integer where 0 means that the row should not be indented

col_wise_args
a named list containing collections (e.g. vectors or lists) with data elements for each column of the resulting table. The data elements are then passed to the named argument FUN corresponding to the element name of the outer list. Hence, the length and order of each collection must match the levels in col_by. See examples.

Value

an rtable object

Examples

rtabulate(x = iris$Species)

#>                   All
#> ---------------------
#> setosa            50 
#> versicolor        50 
#> virginica         50 

rtabulate(x = iris$Species, useNA = "always")

#>                   All
#> ---------------------
#> setosa            50 
#> versicolor        50 
#> virginica         50 
#> <NA>               0 

rtabulate(x = factor(c("a", "a", NA, "b")), useNA = "ifany")

#>             All
#> ---------------
#> a            2 
#> b            1 
#> <NA>         1 

rtabulate(x = iris$Species, by_all("sum"))

#>                   sum
#> ---------------------
#> setosa            50 
#> versicolor        50 
#> virginica         50 

sl5 <- factor(iris$Sepal.Length > 5, levels = c(TRUE, FALSE),
   labels = c("S.L > 5", "S.L <= 5"))

rtabulate(iris$Species, col_by = sl5)

#>                   S.L > 5         S.L <= 5
#> ------------------------------------------
#> setosa               22              28   
#> versicolor           47              3    
#> virginica            49              1    

rtabulate(sl5, iris$Species)

#>                   setosa          versicolor        virginica 
#> --------------------------------------------------------------
#> S.L > 5             22                47                49    
#> S.L <= 5            28                3                 1     

rtabulate(iris$Species, col_by = sl5,
   FUN = function(cell_data, N) {
     if (length(cell_data) > 10) {
        length(cell_data) * c(1, 1/N)
     } else {
        rcell("-", format = "xx")
     }
   },
   format = "xx (xx.xx%)",
   col_wise_args = list(N = table(sl5))
)

#>                     S.L > 5           S.L <= 5  
#> ------------------------------------------------
#> setosa            22 (18.64%)        28 (87.5%) 
#> versicolor        47 (39.83%)             -     
#> virginica         49 (41.53%)             -     

rtabulate(x = factor(c("X", "Y"), c("X", "Y")),
          col_by = factor(c("a", "a"), c("a", "b")), FUN = length)

#>          a        b
#> -------------------
#> X        1        0
#> Y        1        0

rtabulate(factor(c("Y", "Y"), c("X", "Y")),
          factor(c("b", "b"), c("a", "b")), length)

#>          a        b
#> -------------------
#> X        0        0
#> Y        0        2

rtabulate(
  x = factor(c("Y", "Y"), c("X", "Y")),
  col_by = factor(c("b", "b"), c("a", "b")),
  FUN = function(x, N) list(length(x), N),
  col_wise_args = list(N = c(1,2))
)

#>           a           b  
#> -------------------------
#> X        0, 1        0, 2
#> Y        0, 1        2, 2