tabulate a logical vector

tabulate a logical vector

# S3 method for logical
rtabulate(x, col_by = no_by("col_1"), FUN = function(x)
  sum(x) * c(1, 1/length(x)), row_data_arg = FALSE,
  format = "xx.xx (xx.xx%)", row.name = "")

Arguments

x
a vecor

col_by
a vector of length nrow(x) that defines which levels in col_by define a column. If data should not be split into columns use the no_by function.

FUN
a function that processes the cell data, if row_data_arg is set to TRUE the a second argument with the row data is passed to FUN

row_data_arg
call FUN with the row data as the second argument

format
if FUN does not return a formatted rcell then the format is applied

row.name
if NULL then the FUN argument is deparsed and used as row.name of the rrow

Value

a slingle rrow with columns according to the levels of col_by

Examples

rtabulate(iris$Species == "setosa")

#>            col_1   
#> -------------------
#>         50 (33.33%)

rtabulate(iris$Species == "setosa", no_by("Species"), row.name = "n (n/N)")

#>                  Species  
#> --------------------------
#> n (n/N)        50 (33.33%)

# default: percentages equal \code{TRUE}
with(iris, rtabulate(Sepal.Length < 5, Species, row.name = "Sepal.Length < 5"))

#>                           setosa          versicolor        virginica 
#> ----------------------------------------------------------------------
#> Sepal.Length < 5         20 (40%)           1 (2%)            1 (2%)  

# precentages with proportion of cell number of \code{TRUE}s to overvall
# number of \code{TRUE}s
with(iris, rtabulate(Sepal.Length < 5, Species, row.name = "Sepal.Length < 5",
  FUN = function(cell_data, row_data) sum(cell_data) * c(1, 1/sum(row_data)),
  row_data_arg = TRUE
))

#>                           setosa           versicolor          virginica 
#> -------------------------------------------------------------------------
#> Sepal.Length < 5        20 (90.91%)         1 (4.55%)          1 (4.55%)