tabulate a numeric vector
Source: R/rtabulate.R

by default the fivenum function is applied to the vectors associated to the cells

# S3 method for numeric
rtabulate(x, col_by = no_by("col_1"), FUN = mean, ...,
  row_data_arg = FALSE, format = NULL, row.name = NULL, indent = 0)

Arguments

x
a vecor

col_by
a factor of length nrow(x) that defines which levels in col_by define a column. If data should not be split into columns use the no_by function.

FUN
a function that processes the cell data, if row_data_arg is set to TRUE the a second argument with the row data is passed to FUN

...
arguments passed to FUN

row_data_arg
call FUN with the row data as the second argument

format
if FUN does not return a formatted rcell then the format is applied

row.name
if NULL then the FUN argument is deparsed and used as row.name of the rrow

indent
non-negative integer where 0 means that the row should not be indented

Value

an rtable project

Examples

rtabulate(iris$Sepal.Length)

#>                  col_1      
#> ----------------------------
#> mean        5.84333333333333

rtabulate(iris$Sepal.Length, col_by = no_by("Sepal.Length"))

#>               Sepal.Length  
#> ----------------------------
#> mean        5.84333333333333

with(iris,  rtabulate(x = Sepal.Length, col_by = Species, row.name = "fivenum"))

#>                  setosa          versicolor        virginica 
#> -------------------------------------------------------------
#> fivenum          5.006             5.936             6.588   

SL <- iris$Sepal.Length
Sp <- iris$Species
rbind(
  rtabulate(SL, Sp, length, row.name = "n"),
  rtabulate(SL, Sp, function(x)c(mean(x), sd(x)), format = "xx.xx (xx.xx)", row.name = "Mean (SD)"),
  rtabulate(SL, Sp, median, row.name = "Median"),
  rtabulate(SL, Sp, range, format = "xx.xx - xx.xx", row.name = "Min - Max")
)

#>                    setosa           versicolor          virginica 
#> ------------------------------------------------------------------
#> n                    50                 50                 50     
#> Mean (SD)        5.01 (0.35)        5.94 (0.52)        6.59 (0.64)
#> Median                5                 5.9                6.5    
#> Min - Max         4.3 - 5.8           4.9 - 7           4.9 - 7.9