rtabulate default for vectors

this method is used for vectors of type logical and numeric

rtabulate_default(x, col_by = no_by("col_1"), FUN = NULL,
  row_data_arg = FALSE, format = NULL, row.name = "", indent = 0)

Arguments

x
a vecor

col_by
a vector of length nrow(x) that defines which levels in col_by define a column. If data should not be split into columns use the no_by function.

FUN
a function that processes the cell data, if row_data_arg is set to TRUE the a second argument with the row data is passed to FUN

row_data_arg
call FUN with the row data as the second argument

format
if FUN does not return a formatted rcell then the format is applied

row.name
string with row name

indent
non-negative integer where 0 means that the row should not be indented

Value

a slingle rrow with columns according to the levels of col_by