Score functions for sorting TableTrees

Score functions for sorting TableTrees

Usage

cont_n_allcols(tt)

cont_n_onecol(j)

Arguments

tt

(TableTree or related class)
a TableTree object representing a populated table.

j

(numeric(1))
index of column used for scoring.

Value

A single numeric value indicating score according to the relevant metric for tt, to be used when sorting.

See also