Translate spl_context to a path to display in error messages

Translate spl_context to a path to display in error messages

Usage

spl_context_to_disp_path(ctx)

Arguments

ctx

(data.frame)
the spl_context data frame where the error occurred.

Value

A character string containing a description of the row path corresponding to ctx.