Declaring a column-split based on levels of a variable

Will generate children for each subset of a categorical variable

split_cols_by(
  lyt,
  var,
  labels_var = var,
  split_label = var,
  split_fun = NULL,
  format = NULL,
  nested = TRUE,
  child_labels = c("default", "visible", "hidden"),
  extra_args = list(),
  ref_group = NULL
)

Arguments

lyt
layout object pre-data used for tabulation

var
string, variable name

labels_var
string, name of variable containing labels to be displayed for the values of var

split_label
string. Label string to be associated with the table generated by the split. Not to be confused with labels assigned to each child (which are based on the data and type of split during tabulation).

split_fun
function/NULL. custom splitting function See custom_split_funs

format
FormatSpec. Format associated with this split. Formats can be declared via strings ("xx.x") or function. In cases such as analyze calls, they can character vectors or lists of functions.

nested
boolean, Add this as a new top-level split (defining a new subtable directly under root). Defaults to FALSE

child_labels
string. One of "default", "visible", "hidden". What should the display behavior be for the labels (ie label rows) of the children of this split. Defaults to "default" which flags the label row as visible only if the child has 0 content rows.

extra_args
list. Extra arguments to be passed to the tabulation function. Element position in thte list corresponds to the children of this split. Named elements in the child-specific lists are ignored if they do not match a formal argument of the ttabulation function.

ref_group
character(1) or NULL. Level of var which should be considered ref_group/reference

Value

A PreDataTableLayouts object suitable for passing to further layouting functions, and to build_table.

Custom Splitting Function Details

User-defined custom split functions can perform any type of computation on the incoming data provided that they meed the contract for generating 'splits' of the incoming data 'based on' the split object.

Split functions are functions that accept:

df

data.frame of incoming data to be split

spl

a Split object. this is largely an internal detail custom functions will not need to worry about, but obj_name(spl), for example, will give the name of the split as it will appear in paths in the resulting table

vals

Any pre-calculated values. If given non-null values, the values returned should match these. Should be NULL in most cases and can likely be ignored

labels

Any pre-calculated value labels. Same as above for values

trim

If TRUE, resulting splits that are empty should be removed

(Optional) .spl_context

a data.frame describing previously performed splits which collectively arrived at df

The function must then output a named list with the following elements:

values

The vector of all values corresponding to the splits of df

datasplit

a list of data.frames representing the groupings of the actual observations from df.

labels

a character vector giving a string label for each value listed in the values element above

(Optional) extras

If present, extra arguments to be passed to summary and analysis functions whenever they are executed on the corresponding element of datasplit or a subset thereof

One way to generate custom splitting functions is to wrap existing split functions and modify either the incoming data before they are called, or their outputs.

Author

Gabriel Becker

Examples

lyt <- basic_table() %>%
  split_cols_by("ARM") %>%
  analyze(c("AGE", "BMRKR2"))

build_table(lyt, ex_adsl)
#>            A: Drug X   B: Placebo   C: Combination
#> --------------------------------------------------
#> AGE                                               
#>   Mean       33.77       35.43          35.43     
#> BMRKR2                                            
#>   LOW         50           45             40      
#>   MEDIUM      37           56             42      
#>   HIGH        47           33             50      

# Let's look at the splits in more detail

l <- basic_table() %>% split_cols_by("ARM")
l
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ARM (lvls) 
#> 
#> Row-Split Structure:
#> ( () ->  () -> ) () 
#> 

# add an analysis (summary)
l2 <- l %>%
    analyze(c("AGE", "COUNTRY"), afun = list_wrap_x(summary) , format = "xx.xx")
l2
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ARM (lvls) 
#> 
#> Row-Split Structure:
#> ( (** multivar analysis **) -> AGE, COUNTRY (** multivar analysis **) -> ) (** multivar analysis **) 
#> 

build_table(l2, DM)
#>             A: Drug X   B: Placebo   C: Combination
#> ---------------------------------------------------
#> AGE                                                
#>   Min.         20           21             22      
#>   1st Qu.      29           29             30      
#>   Median       33           32             33      
#>   Mean        34.91       33.02          34.57     
#>   3rd Qu.      39           37             38      
#>   Max.         60           55             53      
#> COUNTRY                                            
#>   CHN          62           48             69      
#>   USA          13           14             17      
#>   BRA           9           13             7       
#>   PAK           8           8              12      
#>   NGA          10           5              9       
#>   RUS           9           5              6       
#>   JPN           5           8              5       
#>   GBR           2           3              2       
#>   CAN           3           2              2       
#>   CHE           0           0              0       

# By default sequentially adding layouts results in nesting
library(dplyr)
DM_MF <- DM %>% filter(SEX %in% c("M", "F")) %>% mutate(SEX = droplevels(SEX))

l3 <- basic_table() %>% split_cols_by("ARM") %>%
  split_cols_by("SEX") %>%
  analyze(c("AGE", "COUNTRY"), afun = list_wrap_x(summary), format = "xx.xx")
l3
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ARM (lvls) -> SEX (lvls) 
#> 
#> Row-Split Structure:
#> ( (** multivar analysis **) -> AGE, COUNTRY (** multivar analysis **) -> ) (** multivar analysis **) 
#> 

 build_table(l3, DM_MF)
#>               A: Drug X      B: Placebo      C: Combination  
#>               F       M       F       M        F         M   
#> -------------------------------------------------------------
#> AGE                                                          
#>   Min.       20      24      21      21       22        25   
#>   1st Qu.    29      31      29      28       30        29   
#>   Median     32      35      33      31       35        32   
#>   Mean      33.71   36.55   33.84   32.1     34.89     34.28 
#>   3rd Qu.    38     41.5     38     35.75     39        38   
#>   Max.       58      60      55      47       53        53   
#> COUNTRY                                                      
#>   CHN        34      28      29      19       31        38   
#>   USA         8       5       6       8       10         7   
#>   BRA         6       3       6       7        3         4   
#>   PAK         2       6       5       3        5         7   
#>   NGA         6       4       2       3        5         4   
#>   RUS         7       2       1       4        2         4   
#>   JPN         2       3       3       5        4         1   
#>   GBR         2       0       3       0        1         1   
#>   CAN         3       0       1       1        0         2   
#>   CHE         0       0       0       0        0         0   

# nested=TRUE vs not
l4 <- basic_table() %>% split_cols_by("ARM") %>%
 split_rows_by("SEX", split_fun = drop_split_levels) %>%
 split_rows_by("RACE", split_fun = drop_split_levels) %>%
 analyze("AGE")

l4
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ARM (lvls) 
#> 
#> Row-Split Structure:
#> SEX (lvls) -> RACE (lvls) -> AGE (** analysis **) 
#> 
build_table(l4, DM)
#>                               A: Drug X   B: Placebo   C: Combination
#> ---------------------------------------------------------------------
#> F                                                                    
#>   ASIAN                                                              
#>     Mean                        33.55         34            34.9     
#>   BLACK OR AFRICAN AMERICAN                                          
#>     Mean                        33.17       30.58          33.85     
#>   WHITE                                                              
#>     Mean                        35.88       38.57           36.5     
#> M                                                                    
#>   ASIAN                                                              
#>     Mean                        35.03        31.1          34.39     
#>   BLACK OR AFRICAN AMERICAN                                          
#>     Mean                        37.4        32.83          34.14     
#>   WHITE                                                              
#>     Mean                         44         35.29            34      

l5 <- basic_table() %>% split_cols_by("ARM") %>%
 split_rows_by("SEX", split_fun= drop_split_levels) %>%
 analyze("AGE") %>%
 split_rows_by("RACE", nested=FALSE, split_fun = drop_split_levels) %>%
 analyze("AGE")

l5
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ARM (lvls) 
#> 
#> Row-Split Structure:
#> SEX (lvls) -> AGE (** analysis **) 
#> RACE (lvls) -> AGE (** analysis **) 
#> 
build_table(l5, DM)
#>                             A: Drug X   B: Placebo   C: Combination
#> -------------------------------------------------------------------
#> F                                                                  
#>   Mean                        33.71       33.84          34.89     
#> M                                                                  
#>   Mean                        36.55        32.1          34.28     
#> ASIAN                                                              
#>   Mean                        34.2        32.68          34.63     
#> BLACK OR AFRICAN AMERICAN                                          
#>   Mean                        34.68       31.71            34      
#> WHITE                                                              
#>   Mean                        39.36       36.93          35.11