Associate Multiple Variables with Columns

In some cases, the variable to be ultimately analyzed is most naturally defined on a column, not a row basis. When we need columns to reflect different variables entirely, rather than different levels of a single variable, we use split_cols_by_multivar

split_cols_by_multivar(
  lyt,
  vars,
  varlabels = vars,
  varnames = NULL,
  nested = TRUE
)

Arguments

lyt
layout object pre-data used for tabulation

vars
character vector. Multiple variable names.

varlabels
character vector. Labels for vars

varnames
character vector. Names for vars which will appear in pathing. When vars are all unique this will be the variable names. If not, these will be variable names with suffixes as necessary to enforce uniqueness.

nested
boolean, Add this as a new top-level split (defining a new subtable directly under root). Defaults to FALSE

Value

A PreDataTableLayouts object suitable for passing to further layouting functions, and to build_table.

See also

Author

Gabriel Becker

Examples

library(dplyr)
ANL <- DM %>% mutate(value = rnorm(n()), pctdiff = runif(n()))

## toy example where we take the mean of the first variable and the
## count of >.5 for the second.
colfuns <- list(function(x) in_rows(mean = mean(x), .formats = "xx.x"),
                function(x) in_rows("# x > 5" = sum(x > .5), .formats = "xx"))

l <- basic_table() %>%
    split_cols_by("ARM") %>%
    split_cols_by_multivar(c("value", "pctdiff")) %>%
    split_rows_by("RACE", split_label = "ethnicity", split_fun = drop_split_levels) %>%
    summarize_row_groups() %>%
    analyze_colvars(afun = colfuns)

l

#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ARM (lvls) -> value:pctdiff (vars) 
#> 
#> Row-Split Structure:
#> RACE (lvls) -> NA (** col-var analysis **) 
#> 

build_table(l, ANL)

#>                                    A: Drug X                B: Placebo              C: Combination     
#>                               value       pctdiff       value       pctdiff       value       pctdiff  
#> -------------------------------------------------------------------------------------------------------
#> ASIAN                       79 (65.3%)   79 (65.3%)   68 (64.2%)   68 (64.2%)   84 (65.1%)   84 (65.1%)
#>   mean                          0            39          -0.1          31           0            33    
#> BLACK OR AFRICAN AMERICAN   28 (23.1%)   28 (23.1%)   24 (22.6%)   24 (22.6%)   27 (20.9%)   27 (20.9%)
#>   mean                         0.2           11          0.1           10          -0.3          10    
#> WHITE                       14 (11.6%)   14 (11.6%)   14 (13.2%)   14 (13.2%)    18 (14%)     18 (14%) 
#>   mean                         0.3           8           0.4           7           -0.2          9