Add Rows according to levels of a variable

Add Rows according to levels of a variable

split_rows_by(
  lyt,
  var,
  labels_var = var,
  split_label = var,
  split_fun = NULL,
  format = NULL,
  nested = TRUE,
  child_labels = c("default", "visible", "hidden"),
  visible_label = FALSE,
  indent_mod = 0L
)

Arguments

lyt
layout object pre-data used for tabulation

var
string, variable name

labels_var
string, name of variable containing labels to be displayed for the values of var

split_label
string. Label string to be associated with the table generated by the split. Not to be confused with labels assigned to each child (which are based on the data and type of split during tabulation).

split_fun
function/NULL. custom splitting function

format
FormatSpec. Format associated with this split. Formats can be declared via strings ("xx.x") or function. In cases such as analyze calls, they can character vectors or lists of functions.

nested
boolean, Add this as a new top-level split (defining a new subtable directly under root). Defaults to FALSE

child_labels
string. One of "default", "visible", "hidden". What should the display behavior be for the labels (ie label rows) of the children of this split. Defaults to "default" which flags the label row as visible only if the child has 0 content rows.

visible_label
logical(1). Should the split label be visible in the resulting table.

indent_mod
numeric. Modifier for the default indent position for the structure created by this function(subtable, content table, or row) and all of that structure's children. Defaults to 0, which corresponds to the unmodified default behavior.

Value

A PreDataTableLayouts object suitable for passing to further layouting functions, and to build_table.

Note

If var is a factor with empty unobserved levels and labels_var is specified, it must also be a factor with the same number of levels as var. Currently the error that occurs when this is not hte case is not very informative, but that will change in the future.

Author

Gabriel Becker

Examples

l <- basic_table() %>%
    split_cols_by("ARM") %>%
    split_rows_by("RACE", split_fun = drop_split_levels) %>%
    analyze("AGE", mean, var_labels = "Age", format = "xx.xx")

build_table(l, DM)

#>                             A: Drug X   B: Placebo   C: Combination
#> -------------------------------------------------------------------
#> ASIAN                                                              
#>   mean                        34.2        32.68          34.63     
#> BLACK OR AFRICAN AMERICAN                                          
#>   mean                        34.68       31.71            34      
#> WHITE                                                              
#>   mean                        39.36       36.93          35.11     

basic_table() %>%
    split_cols_by("ARM") %>%
    split_rows_by("RACE") %>%
    analyze("AGE", mean, var_labels = "Age", format = "xx.xx") %>%
    build_table(DM)

#>                                             A: Drug X   B: Placebo   C: Combination
#> -----------------------------------------------------------------------------------
#> ASIAN                                                                              
#>   mean                                        34.2        32.68          34.63     
#> BLACK OR AFRICAN AMERICAN                                                          
#>   mean                                        34.68       31.71            34      
#> WHITE                                                                              
#>   mean                                        39.36       36.93          35.11     
#> AMERICAN INDIAN OR ALASKA NATIVE                                                   
#>   mean                                         NaN         NaN            NaN      
#> MULTIPLE                                                                           
#>   mean                                         NaN         NaN            NaN      
#> NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER                                          
#>   mean                                         NaN         NaN            NaN      
#> OTHER                                                                              
#>   mean                                         NaN         NaN            NaN      
#> UNKNOWN                                                                            
#>   mean                                         NaN         NaN            NaN      

l <- basic_table() %>%
    split_cols_by("ARM") %>%
    split_cols_by("SEX") %>%
    summarize_row_groups(label_fstr = "Overall (N)") %>%
    split_rows_by("RACE", split_label = "Ethnicity", labels_var = "ethn_lab",
                  split_fun = drop_split_levels) %>%
    summarize_row_groups("RACE", label_fstr = "%s (n)") %>%
    analyze("AGE", var_labels = "Age", afun = mean, format = "xx.xx")

l

#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ARM (lvls) -> SEX (lvls) 
#> 
#> Row-Split Structure:
#> RACE (lvls) -> AGE (** analysis **) 
#> 

library(dplyr)
DM2 <- DM %>%
    filter(SEX %in% c("M", "F")) %>%
    mutate(
        SEX = droplevels(SEX),
        gender_lab = c("F" = "Female", "M" = "Male",
                       "U" = "Unknown", "UNDIFFERENTIATED" = "Undifferentiated")[SEX],
        ethn_lab = c(
            "ASIAN" = "Asian",
            "BLACK OR AFRICAN AMERICAN" = "Black or African American",
            "WHITE" = "White",
            "AMERICAN INDIAN OR ALASKA NATIVE" = "American Indian or Alaska Native",
             "MULTIPLE" = "Multiple",
             "NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER" =
                 "Native Hawaiian or Other Pacific Islander",
             "OTHER" = "Other", "UNKNOWN" = "Unknown"
        )[RACE]
    )

build_table(l, DM2)

#>                                          A: Drug X                B: Placebo             C: Combination     
#>                                       F            M            F            M           F            M     
#> ------------------------------------------------------------------------------------------------------------
#> Overall (N)                       70 (100%)    51 (100%)    56 (100%)    50 (100%)   61 (100%)    68 (100%) 
#>   Asian (n)                       44 (62.9%)   35 (68.6%)   37 (66.1%)   31 (62%)    40 (65.6%)   44 (64.7%)
#>     mean                            33.55        35.03          34         31.1         34.9        34.39   
#>   Black or African American (n)   18 (25.7%)   10 (19.6%)   12 (21.4%)   12 (24%)    13 (21.3%)   14 (20.6%)
#>     mean                            33.17         37.4        30.58        32.83       33.85        34.14   
#>   White (n)                       8 (11.4%)    6 (11.8%)    7 (12.5%)     7 (14%)    8 (13.1%)    10 (14.7%)
#>     mean                            35.88          44         38.57        35.29        36.5          34