TableTree classes

TableTree classes

Table Constructors and Classes

ElementaryTable(
  kids = list(),
  name = "",
  lev = 1L,
  label = "",
  labelrow = LabelRow(lev = lev, label = label, vis = !isTRUE(iscontent) &&
    !is.na(label) && nzchar(label)),
  rspans = data.frame(),
  cinfo = NULL,
  iscontent = NA,
  var = NA_character_,
  format = NULL,
  indent_mod = 0L,
  title = "",
  subtitles = character(),
  main_footer = character(),
  prov_footer = character(),
  hsep = default_hsep()
)

TableTree(
  kids = list(),
  name = if (!is.na(var)) var else "",
  cont = EmptyElTable,
  lev = 1L,
  label = name,
  labelrow = LabelRow(lev = lev, label = label, vis = nrow(cont) == 0 && !is.na(label)
    && nzchar(label)),
  rspans = data.frame(),
  iscontent = NA,
  var = NA_character_,
  cinfo = NULL,
  format = NULL,
  indent_mod = 0L,
  title = "",
  subtitles = character(),
  main_footer = character(),
  prov_footer = character(),
  page_title = NA_character_,
  hsep = default_hsep()
)

Arguments

kids

list. List of direct children.

name

character(1). Name of the split/table/row being creatted. Defaults to same as the corresponding llabel, but is not required to be.

lev

integer. Nesting level (roughly, indentation level in practical terms).

label

character(1). A label (not to be confused with the name) for the object/structure.

labelrow

LabelRow. The LabelRow object to assign to this Table. Consructed from label by default if not specified.

rspans

data.frame. Currently stored but otherwise ignored.

cinfo

InstantiatedColumnInfo (or NULL). Column structure for the object being created.

iscontent

logical. Is the TableTree/ElementaryTable being constructed the content table for another TableTree.

var

string, variable name

format

FormatSpec. Format associated with this split. Formats can be declared via strings ("xx.x") or function. In cases such as analyze calls, they can character vectors or lists of functions.

indent_mod

numeric. Modifier for the default indent position for the structure created by this function(subtable, content table, or row) and all of that structure's children. Defaults to 0, which corresponds to the unmodified default behavior.

title

character(1). Main title. Ignored for subtables.

subtitles

character. Subtitles. Ignored for subtables.

main_footer

character. Main global (non-referential) footer materials.

prov_footer

character. Provenance-related global footer materials. Generally should not be modified by hand.

hsep

character(1). Set of character(s) to be repeated as the separator between the header and body of the table when rendered as text. Defaults to a connected horrizontal line (unicode 2014) in locals that use a UTF charset, and to - elsewhere (with a once per session warning).

cont

ElementaryTable. Content table.

page_title

character. Page specific title(s).

Value

A formal object representing a populated table.

Author

Gabriel Becker

Gabriel Becker