Listing Shenanigans

Experimental. Not for production work. I mean it.

as_listing(
  df,
  cols = key_cols,
  key_cols = names(df)[1],
  title = NULL,
  footer = NULL
)

as_keycol(vec)

is_keycol(vec)

get_keycols(df)

listing_matrix_form(df)

# S4 method for MatrixPrintForm
main_title(obj)

# S4 method for MatrixPrintForm
subtitles(obj)

# S4 method for MatrixPrintForm
main_footer(obj)

# S4 method for MatrixPrintForm
prov_footer(obj)

listing_dispcols(df)

add_listing_dispcol(df, new)

listing_dispcols(df) <- value

add_listing_col(df, name, fun = NULL, format = NULL)

Arguments

df
listing_df. The listing to modify.

cols
character. Names of columns (including but not limited to key columns) which should be displayed when the listing is rendered.

key_cols
character. Names of columns which should be treated as key columns when rendering the listing.

title
character or NULL. Title for the listing. Currently ignored.

footer
character or NULL. Footer for the listing. Currently ignored.

vec
any. A column vector from a listing_df

obj
ANY. The object for the accessor to access or modify

new
character. Names of columns to be added to the set of display columns.

value
The new value

name
character(1). Name of the existing or new column to be displayed when the listing is rendered

fun
function or NULL. A function which accepts df and returns the vector for a new column, which is added to df as name, or NULL if marking an existing column as a listing column

format
FormatSpec. A format specification (format string, function, or sprintf format) for use when displaying the column during rendering.

Value

A listing_df object, sorted by the key columns.

df, with name created (if necessary) and marked for display during rendering.

Examples

dat <- ex_adae

lsting <- as_listing(dat[1:25,], key_cols = c("USUBJID", "AESOC")) %>%
    add_listing_col("AETOXGR") %>%
    add_listing_col("BMRKR1", format = "xx.x") %>%
    add_listing_col("AESER / AREL", fun = function(df) paste(df$AESER, df$AREL, sep = " / "))
#> sorting incoming data by key columns

mat <- listing_matrix_form(lsting)

cat(toString(mat))
#> Unique Subject Identifier   Primary System Organ Class   Analysis Toxicity Grade   Continous Level Biomarker 1   AESER / AREL
#> -----------------------------------------------------------------------------------------------------------------------------
#> AB12345-BRA-1-id-134        cl A                                    2                          6.5                  Y / N    
#>                                                                     2                          6.5                  Y / N    
#>                             cl B                                    3                          6.5                  N / N    
#>                             cl D                                    3                          6.5                  N / N    
#> AB12345-BRA-1-id-141        cl A                                    1                          7.5                  N / N    
#>                                                                     2                          7.5                  Y / N    
#>                                                                     1                          7.5                  N / N    
#>                             cl B                                    3                          7.5                  N / N    
#>                             cl D                                    1                          7.5                  N / Y    
#>                                                                     5                          7.5                  Y / N    
#> AB12345-BRA-1-id-236        cl B                                    5                          7.7                  N / Y    
#>                                                                     5                          7.7                  N / Y    
#>                                                                     5                          7.7                  N / Y    
#> AB12345-BRA-1-id-265        cl C                                    2                         10.3                  N / Y    
#>                                                                     4                         10.3                  N / Y    
#>                             cl D                                    3                         10.3                  N / N    
#>                                                                     5                         10.3                  Y / N    
#> AB12345-BRA-1-id-42         cl A                                    2                          2.3                  Y / N    
#>                                                                     2                          2.3                  Y / N    
#>                                                                     2                          2.3                  Y / N    
#>                             cl B                                    1                          2.3                  Y / N    
#>                                                                     5                          2.3                  N / Y    
#>                             cl C                                    2                          2.3                  N / Y    
#>                                                                     2                          2.3                  N / Y    
#>                             cl D                                    5                          2.3                  Y / N