Split between two or more different variables

Split between two or more different variables

Usage

MultiVarSplit(
  vars,
  split_label = "",
  varlabels = NULL,
  varnames = NULL,
  cfun = NULL,
  cformat = NULL,
  cna_str = NA_character_,
  split_format = NULL,
  split_na_str = NA_character_,
  split_name = "multivars",
  child_labels = c("default", "visible", "hidden"),
  extra_args = list(),
  indent_mod = 0L,
  cindent_mod = 0L,
  cvar = "",
  cextra_args = list(),
  label_pos = "visible",
  split_fun = NULL,
  page_prefix = NA_character_,
  section_div = NA_character_
)

Arguments

vars

character vector. Multiple variable names.

split_label

string. Label string to be associated with the table generated by the split. Not to be confused with labels assigned to each child (which are based on the data and type of split during tabulation).

varlabels

character vector. Labels for vars

varnames

character vector. Names for vars which will appear in pathing. When vars are all unique this will be the variable names. If not, these will be variable names with suffixes as necessary to enforce uniqueness.

cfun

list/function/NULL. tabulation function(s) for creating content rows. Must accept x or df as first parameter. Must accept labelstr as the second argument. Can optionally accept all optional arguments accepted by analysis functions. See analyze.

cformat

format spec. Format for content rows

cna_str

character. NA string for use with cformat for content table.

split_format

FormatSpec. Default format associated with the split being created.

split_na_str

character. NA string vector for use with split_format.

split_name

string. Name associated with this split (for pathing, etc)

child_labels

string. One of "default", "visible", "hidden". What should the display behavior be for the labels (ie label rows) of the children of this split. Defaults to "default" which flags the label row as visible only if the child has 0 content rows.

extra_args

list. Extra arguments to be passed to the tabulation function. Element position in the list corresponds to the children of this split. Named elements in the child-specific lists are ignored if they do not match a formal argument of the tabulation function.

indent_mod

numeric. Modifier for the default indent position for the structure created by this function(subtable, content table, or row) and all of that structure's children. Defaults to 0, which corresponds to the unmodified default behavior.

cindent_mod

numeric(1). The indent modifier for the content tables generated by this split.

cvar

character(1). The variable, if any, which the content function should accept. Defaults to NA.

cextra_args

list. Extra arguments to be passed to the content function when tabulating row group summaries.

label_pos

character(1). Location the variable label should be displayed, Accepts hidden (default for non-analyze row splits), visible, topleft, and - for analyze splits only - default. For analyze calls, default indicates that the variable should be visible if and only if multiple variables are analyzed at the same level of nesting.

split_fun

function/NULL. custom splitting function See custom_split_funs

page_prefix

character(1). Prefix, to be appended with the split value, when forcing pagination between the children of this split/table

section_div

character(1). String which should be repeated as a section divider after each group defined by this split instruction, or NA_character_ (the default) for no section divider.

Value

A MultiVarSplit object.

Author

Gabriel Becker