Split on levels within a variable

Split on levels within a variable

VarLevelSplit(
  var,
  split_label,
  labels_var = NULL,
  cfun = NULL,
  cformat = NULL,
  split_fun = NULL,
  split_format = NULL,
  valorder = NULL,
  split_name = var,
  child_labels = c("default", "visible", "hidden"),
  extra_args = list(),
  indent_mod = 0L,
  label_pos = c("topleft", "hidden", "visible"),
  cindent_mod = 0L,
  cvar = "",
  cextra_args = list(),
  page_prefix = NA_character_
)

VarLevWBaselineSplit(
  var,
  ref_group,
  labels_var = var,
  split_label,
  split_fun = NULL,
  label_fstr = "%s - %s",
  cfun = NULL,
  cformat = NULL,
  cvar = "",
  split_format = NULL,
  valorder = NULL,
  split_name = var,
  extra_args = list()
)

Arguments

var

string, variable name

split_label

string. Label string to be associated with the table generated by the split. Not to be confused with labels assigned to each child (which are based on the data and type of split during tabulation).

labels_var

string, name of variable containing labels to be displayed for the values of var

cfun

list/function/NULL. tabulation function(s) for creating content rows. Must accept x or df as first parameter. Must accept labelstr as the second argument. Can optionally accept all optional arguments accepted by analysis functions. See analyze.

cformat

format spec. Format for content rows

split_fun

function/NULL. custom splitting function See custom_split_funs

split_format

FormatSpec. Default format associated with the split being created.

valorder

character vector. Order that the split children should appear in resulting table.

split_name

string. Name associiated with this split (for pathing, etc)

child_labels

string. One of "default", "visible", "hidden". What should the display behavior be for the labels (ie label rows) of the children of this split. Defaults to "default" which flags the label row as visible only if the child has 0 content rows.

extra_args

list. Extra arguments to be passed to the tabulation function. Element position in thte list corresponds to the children of this split. Named elements in the child-specific lists are ignored if they do not match a formal argument of the ttabulation function.

indent_mod

numeric. Modifier for the default indent position for the structure created by this function(subtable, content table, or row) and all of that structure's children. Defaults to 0, which corresponds to the unmodified default behavior.

label_pos

character(1). Location the variable label should be displayed, Accepts hidden (default for non-analyze row splits), visible, topleft, and - for analyze splits only - default. For analyze calls, default indicates that the variable should be visible if and only if multiple variables are analyzed at the same level of nesting.

cindent_mod

numeric(1). The indent modifier for the content tables generated by this split.

cvar

character(1). The variable, if any, which the content function should accept. Defaults to NA.

cextra_args

list. Extra arguments to be passed to the content function when tabulating row group summaries.

page_prefix

character(1). Prefix, to be appended with the split value, when forcing pagination between the children of this split/table

ref_group

character. Value of var to be taken as the ref_group/control to be compared against.

label_fstr

string. An sprintf style format string containing. For non-comparison splits, it can contain up to one "%s" which takes the current split value and generates the row/column label. Comparison-based splits it can contain up to two "%s".

Value

a VarLevelSplit object.

Author

Gabriel Becker