Constructor Arg Conventions

Constructor Arg Conventions

constr_args(
  kids,
  cont,
  lev,
  iscontent,
  cinfo,
  labelrow,
  vals,
  cspan,
  visible_label,
  cindent_mod,
  cvar,
  label,
  cextra_args,
  child_names
)

Arguments

kids
list. List of direct children.

cont
ElementaryTable. Content table.

lev
integer. Nesting level (roughly, indentation level in practical terms).

iscontent
logical. Is the TableTree/ElementaryTable being constructed the content table for another TableTree.

cinfo
InstantiatedColumnInfo (or NULL). Column structure for the object being created.

labelrow
LabelRow. The LabelRow object to assign to this Table. Consructed from label by default if not specified.

vals
list. cell values for the row

cspan
integer. Column span. 1 indicates no spanning.

visible_label
logical(1). Should the split label be visible in the resulting table.

cindent_mod
numeric(1). The indent modifier for the content tables generated by this split.

cvar
character(1). The variable, if any, which the content function should accept. Defaults to NA.

label
character(1). A label (not to be confused with the name) for the object/structure.

cextra_args
list. Extra arguments to be passed to the content function when tabulating row group summaries.

child_names
character. Names to be given to the sub splits contained by a compound split (typically a AnalyzeMultiVars split object).

See also