Constructor argument conventions

Constructor argument conventions

Usage

constr_args(
  kids,
  cont,
  lev,
  iscontent,
  cinfo,
  labelrow,
  vals,
  cspan,
  label_pos,
  cindent_mod,
  cvar,
  label,
  cextra_args,
  child_names,
  title,
  subtitles,
  main_footer,
  prov_footer,
  footnotes,
  page_title,
  page_prefix,
  section_div,
  trailing_section_div,
  split_na_str,
  cna_str,
  inset,
  table_inset,
  header_section_div
)

Arguments

kids

(list)
list of direct children.

cont

(ElementaryTable)
content table.

lev

(integer(1))
nesting level (roughly, indentation level in practical terms).

iscontent

(flag)
whether the TableTree/ElementaryTable is being constructed as the content table for another TableTree.

cinfo

(InstantiatedColumnInfo or NULL)
column structure for the object being created.

labelrow

(LabelRow)
the LabelRow object to assign to the table. Constructed from label by default if not specified.

vals

(list)
cell values for the row.

cspan

(integer)
column span. 1 indicates no spanning.

label_pos

(string)
location where the variable label should be displayed. Accepts "hidden" (default for non-analyze row splits), "visible", "topleft", and "default" (for analyze splits only). For analyze calls, "default" indicates that the variable should be visible if and only if multiple variables are analyzed at the same level of nesting.

cindent_mod

(numeric(1))
the indent modifier for the content tables generated by this split.

cvar

(string)
the variable, if any, that the content function should accept. Defaults to NA.

label

(string)
a label (not to be confused with the name) for the object/structure.

cextra_args

(list)
extra arguments to be passed to the content function when tabulating row group summaries.

child_names

(character)
names to be given to the subsplits contained by a compound split (typically an AnalyzeMultiVars split object).

title

(string)
single string to use as main title (formatters::main_title()). Ignored for subtables.

subtitles

(character)
a vector of strings to use as subtitles (formatters::subtitles()), where every element is printed on a separate line. Ignored for subtables.

main_footer

(character)
a vector of strings to use as main global (non-referential) footer materials (formatters::main_footer()), where every element is printed on a separate line.

prov_footer

(character)
a vector of strings to use as provenance-related global footer materials (formatters::prov_footer()), where every element is printed on a separate line.

footnotes

(list or NULL)
referential footnotes to be applied at current level. In post-processing, this can be achieved with fnotes_at_path<-.

page_title

(character)
page-specific title(s).

page_prefix

(string)
prefix to be appended with the split value when forcing pagination between the children of a split/table.

section_div

(string)
string which should be repeated as a section divider after each group defined by this split instruction, or NA_character_ (the default) for no section divider.

trailing_section_div

(string)
string which will be used as a section divider after the printing of the last row contained in this (sub)table, unless that row is also the last table row to be printed overall, or NA_character_ for none (the default). When generated via layouting, this would correspond to the section_div of the split under which this table represents a single facet.

split_na_str

(character)
NA string vector for use with split_format.

cna_str

(character)
NA string for use with cformat for content table.

inset

(numeric(1))
number of spaces to inset the table header, table body, referential footnotes, and main_footer, as compared to alignment of title, subtitle, and provenance footer. Defaults to 0 (no inset).

table_inset

(numeric(1))
number of spaces to inset the table header, table body, referential footnotes, and main footer, as compared to alignment of title, subtitles, and provenance footer. Defaults to 0 (no inset).

header_section_div

(string)
string which will be used to divide the header from the table. See header_section_div() for the associated getter and setter. Please consider changing last element of section_div() when concatenating tables that require a divider between them.

Value

No return value.

See also