Constructor Arg Conventions

Constructor Arg Conventions

Usage

constr_args(
  kids,
  cont,
  lev,
  iscontent,
  cinfo,
  labelrow,
  vals,
  cspan,
  label_pos,
  cindent_mod,
  cvar,
  label,
  cextra_args,
  child_names,
  title,
  subtitles,
  main_footer,
  prov_footer,
  footnotes,
  page_title,
  page_prefix,
  section_div,
  trailing_sep,
  split_na_str,
  cna_str,
  inset,
  table_inset
)

Arguments

kids

list. List of direct children.

cont

ElementaryTable. Content table.

lev

integer. Nesting level (roughly, indentation level in practical terms).

iscontent

logical. Is the TableTree/ElementaryTable being constructed the content table for another TableTree.

cinfo

InstantiatedColumnInfo (or NULL). Column structure for the object being created.

labelrow

LabelRow. The LabelRow object to assign to this Table. Constructed from label by default if not specified.

vals

list. cell values for the row

cspan

integer. Column span. 1 indicates no spanning.

label_pos

character(1). Location the variable label should be displayed, Accepts hidden (default for non-analyze row splits), visible, topleft, and - for analyze splits only - default. For analyze calls, default indicates that the variable should be visible if and only if multiple variables are analyzed at the same level of nesting.

cindent_mod

numeric(1). The indent modifier for the content tables generated by this split.

cvar

character(1). The variable, if any, which the content function should accept. Defaults to NA.

label

character(1). A label (not to be confused with the name) for the object/structure.

cextra_args

list. Extra arguments to be passed to the content function when tabulating row group summaries.

child_names

character. Names to be given to the sub splits contained by a compound split (typically a AnalyzeMultiVars split object).

title

character(1). Main title. Ignored for subtables.

subtitles

character. Subtitles. Ignored for subtables.

main_footer

character. Main global (non-referential) footer materials.

prov_footer

character. Provenance-related global footer materials. Generally should not be modified by hand.

footnotes

list or NULL. Referential footnotes to be applied at current level

page_title

character. Page specific title(s).

page_prefix

character(1). Prefix, to be appended with the split value, when forcing pagination between the children of this split/table

section_div

character(1). String which should be repeated as a section divider after each group defined by this split instruction, or NA_character_ (the default) for no section divider.

trailing_sep

character(1). String which will be used as a section divider after the printing of the last row contained in this (sub)-table, unless that row is also the last table row to be printed overall, or NA_character_ for none (the default). When generated via layouting, this would correspond to the section_div of the split under which this table represents a single facet.

split_na_str

character. NA string vector for use with split_format.

cna_str

character. NA string for use with cformat for content table.

inset

numeric(1). Number of spaces to inset the table header, table body, referential footnotes, and main_footer, as compared to alignment of title, subtitle, and provenance footer. Defaults to 0 (no inset).

table_inset

numeric(1). Number of spaces to inset the table header, table body, referential footnotes, and main_footer, as compared to alignment of title, subtitle, and provenance footer. Defaults to 0 (no inset).

Value

NULL (this is an argument template dummy function)

See also