Methods for generics in the formatters package

See the formatters documentation for descriptions of these generics.

Usage

# S4 method for VNodeInfo
obj_name(obj)

# S4 method for Split
obj_name(obj)

# S4 method for VNodeInfo
obj_name(obj) <- value

# S4 method for Split
obj_name(obj) <- value

# S4 method for Split
obj_label(obj)

# S4 method for TableRow
obj_label(obj)

# S4 method for VTableTree
obj_label(obj)

# S4 method for ValueWrapper
obj_label(obj)

# S4 method for Split
obj_label(obj) <- value

# S4 method for TableRow
obj_label(obj) <- value

# S4 method for ValueWrapper
obj_label(obj) <- value

# S4 method for VTableTree
obj_label(obj) <- value

# S4 method for VTableNodeInfo
obj_format(obj)

# S4 method for CellValue
obj_format(obj)

# S4 method for Split
obj_format(obj)

# S4 method for VTableNodeInfo
obj_format(obj) <- value

# S4 method for Split
obj_format(obj) <- value

# S4 method for CellValue
obj_format(obj) <- value

# S4 method for Split
obj_na_str(obj)

# S4 method for VTitleFooter
main_title(obj)

# S4 method for VTitleFooter
main_title(obj) <- value

# S4 method for TableRow
main_title(obj)

# S4 method for VTitleFooter
subtitles(obj)

# S4 method for VTitleFooter
subtitles(obj) <- value

# S4 method for TableRow
subtitles(obj)

# S4 method for VTitleFooter
main_footer(obj)

# S4 method for VTitleFooter
main_footer(obj) <- value

# S4 method for TableRow
main_footer(obj)

# S4 method for VTitleFooter
prov_footer(obj)

# S4 method for VTitleFooter
prov_footer(obj) <- value

# S4 method for TableRow
prov_footer(obj)

# S4 method for VTableNodeInfo
table_inset(obj)

# S4 method for PreDataTableLayouts
table_inset(obj)

# S4 method for VTableNodeInfo
table_inset(obj) <- value

# S4 method for PreDataTableLayouts
table_inset(obj) <- value

# S4 method for InstantiatedColumnInfo
table_inset(obj) <- value

# S4 method for TableRow
nlines(x, colwidths = NULL, max_width = NULL)

# S4 method for LabelRow
nlines(x, colwidths = NULL, max_width = NULL)

# S4 method for RefFootnote
nlines(x, colwidths = NULL, max_width = NULL)

# S4 method for InstantiatedColumnInfo
nlines(x, colwidths = NULL, max_width = NULL)

# S4 method for VTableTree
make_row_df(
  tt,
  colwidths = NULL,
  visible_only = TRUE,
  rownum = 0,
  indent = 0L,
  path = character(),
  incontent = FALSE,
  repr_ext = 0L,
  repr_inds = integer(),
  sibpos = NA_integer_,
  nsibs = NA_integer_,
  max_width = NULL
)

# S4 method for TableRow
make_row_df(
  tt,
  colwidths = NULL,
  visible_only = TRUE,
  rownum = 0,
  indent = 0L,
  path = "root",
  incontent = FALSE,
  repr_ext = 0L,
  repr_inds = integer(),
  sibpos = NA_integer_,
  nsibs = NA_integer_,
  max_width = NULL
)

# S4 method for LabelRow
make_row_df(
  tt,
  colwidths = NULL,
  visible_only = TRUE,
  rownum = 0,
  indent = 0L,
  path = "root",
  incontent = FALSE,
  repr_ext = 0L,
  repr_inds = integer(),
  sibpos = NA_integer_,
  nsibs = NA_integer_,
  max_width = NULL
)

Arguments

obj

ANY. The object for the accessor to access or modify

value

The new value

x

An object

colwidths

numeric vector. Column widths for use with vertical pagination.

max_width

numeric(1). Width strings should be wrapped to when determining how many lines they require.

tt

TableTree (or related class). A TableTree object representing a populated table.

visible_only

logical(1). Should only visible aspects of the table structure be reflected in this summary. Defaults to TRUE. May not be supported by all methods.

rownum

numeric(1). Internal detail do not set manually.

indent

integer(1). Internal detail do not set manually.

path

character. A vector path for a position within the structure of a tabletree. Each element represents a subsequent choice amongst the children of the previous choice.

incontent

logical(1). Internal detail do not set manually.

repr_ext

integer(1). Internal detail do not set manually.

repr_inds

integer. Internal detail do not set manually.

sibpos

integer(1). Internal detail do not set manually.

nsibs

integer(1). Internal detail do not set manually.

Value

for getters, the current value of the component being accessed on obj, for setters, a modified copy of obj with the new value.

Details

When visible_only is TRUE (the default), methods should return a data.frame with exactly one row per visible row in the table-like object. This is useful when reasoning about how a table will print, but does not reflect the full pathing space of the structure (though the paths which are given will all work as is).

If supported, when visible_only is FALSE, every structural element of the table (in row-space) will be reflected in the returned data.frame, meaning the full pathing-space will be represented but some rows in the layout summary will not represent printed rows in the table as it is displayed.

Most arguments beyond tt and visible_only are present so that make_row_df methods can call make_row_df recursively and retain information, and should not be set during a top-level call

Note

the technically present root tree node is excluded from the summary returned by both make_row_df and make_col_df (see rtables::make_col_df), as it is simply the row/column structure of tt and thus not useful for pathing or pagination.