List with valid rcell formats labels grouped by 1d and 2d

Currently valid format lables can not be added dynamically. Format functions must be used for special cases

list_rcell_format_labels()

Examples

list_rcell_format_labels()

#> $`1d`
#>  [1] "xx"      "xx."     "xx.x"    "xx.xx"   "xx.xxx"  "xx.xxxx" "xx%"    
#>  [8] "xx.x%"   "xx.xx%"  "xx.xxx%" "(N=xx)"  ">999.9" 
#> 
#> $`2d`
#>  [1] "xx / xx"            "xx. / xx."          "xx.x / xx.x"       
#>  [4] "xx.xx / xx.xx"      "xx (xx%)"           "xx (xx.%)"         
#>  [7] "xx (xx.x%)"         "xx (xx.xx%)"        "xx. (xx.%)"        
#> [10] "xx.x (xx.x%)"       "xx.xx (xx.xx%)"     "(xx, xx)"          
#> [13] "(xx., xx.)"         "(xx.x, xx.x)"       "(xx.xx, xx.xx)"    
#> [16] "(xx.xxx, xx.xxx)"   "(xx.xxxx, xx.xxxx)" "xx - xx"           
#> [19] "xx.x - xx.x"        "xx.xx - xx.xx"      "xx.x (xx.x)"       
#> [22] "xx.xx (xx.xx)"      "xx.x, xx.x"         "xx.x to xx.x"      
#> 
#> attr(,"info")
#> [1] "xx does not modify the element, and xx. rounds a number to 0 digits"