Layouting Function Arg Conventions

Layouting Function Arg Conventions

Usage

lyt_args(
  lyt,
  var,
  vars,
  label,
  labels_var,
  varlabels,
  varnames,
  split_format,
  split_na_str,
  nested,
  format,
  cfun,
  cformat,
  cna_str,
  split_fun,
  split_name,
  split_label,
  afun,
  inclNAs,
  valorder,
  ref_group,
  compfun,
  label_fstr,
  child_labels,
  extra_args,
  name,
  cuts,
  cutlabels,
  cutfun,
  cutlabelfun,
  cumulative,
  indent_mod,
  show_labels,
  label_pos,
  var_labels,
  cvar,
  table_names,
  topleft,
  align,
  page_by,
  page_prefix,
  format_na_str,
  section_div,
  na_str
)

Arguments

lyt

layout object pre-data used for tabulation

var

string, variable name

vars

character vector. Multiple variable names.

label

character(1). A label (not to be confused with the name) for the object/structure.

labels_var

string, name of variable containing labels to be displayed for the values of var

varlabels

character vector. Labels for vars

varnames

character vector. Names for vars which will appear in pathing. When vars are all unique this will be the variable names. If not, these will be variable names with suffixes as necessary to enforce uniqueness.

split_format

FormatSpec. Default format associated with the split being created.

split_na_str

character. NA string vector for use with split_format.

nested

boolean. Should this layout instruction be applied within the existing layout structure if possible (TRUE, the default) or as a new top-level element (`FALSE). Ignored if it would nest a split underneath analyses, which is not allowed.

format

FormatSpec. Format associated with this split. Formats can be declared via strings ("xx.x") or function. In cases such as analyze calls, they can character vectors or lists of functions.

cfun

list/function/NULL. tabulation function(s) for creating content rows. Must accept x or df as first parameter. Must accept labelstr as the second argument. Can optionally accept all optional arguments accepted by analysis functions. See analyze.

cformat

format spec. Format for content rows

cna_str

character. NA string for use with cformat for content table.

split_fun

function/NULL. custom splitting function See custom_split_funs

split_name

string. Name associated with this split (for pathing, etc)

split_label

string. Label string to be associated with the table generated by the split. Not to be confused with labels assigned to each child (which are based on the data and type of split during tabulation).

afun

function. Analysis function, must take x or df as its first parameter. Can optionally take other parameters which will be populated by the tabulation framework. See Details in analyze.

inclNAs

boolean. Should observations with NA in the var variable(s) be included when performing this analysis. Defaults to FALSE

valorder

character vector. Order that the split children should appear in resulting table.

ref_group

character. Value of var to be taken as the ref_group/control to be compared against.

compfun

function/string. The comparison function which accepts the analysis function outputs for two different partitions and returns a single value. Defaults to subtraction. If a string, taken as the name of a function.

label_fstr

string. An sprintf style format string containing. For non-comparison splits, it can contain up to one "%s" which takes the current split value and generates the row/column label. Comparison-based splits it can contain up to two "%s".

child_labels

string. One of "default", "visible", "hidden". What should the display behavior be for the labels (ie label rows) of the children of this split. Defaults to "default" which flags the label row as visible only if the child has 0 content rows.

extra_args

list. Extra arguments to be passed to the tabulation function. Element position in the list corresponds to the children of this split. Named elements in the child-specific lists are ignored if they do not match a formal argument of the tabulation function.

name

character(1). Name of the split/table/row being created. Defaults to same as the corresponding label, but is not required to be.

cuts

numeric. Cuts to use

cutlabels

character (or NULL). Labels for the cuts

cutfun

function. Function which accepts the full vector of var values and returns cut points to be used (via cut) when splitting data during tabulation

cutlabelfun

function. Function which returns either labels for the cuts or NULL when passed the return value of cutfun

cumulative

logical. Should the cuts be treated as cumulative. Defaults to FALSE

indent_mod

numeric. Modifier for the default indent position for the structure created by this function(subtable, content table, or row) and all of that structure's children. Defaults to 0, which corresponds to the unmodified default behavior.

show_labels

character(1). Should the variable labels for corresponding to the variable(s) in vars be visible in the resulting table.

label_pos

character(1). Location the variable label should be displayed, Accepts hidden (default for non-analyze row splits), visible, topleft, and - for analyze splits only - default. For analyze calls, default indicates that the variable should be visible if and only if multiple variables are analyzed at the same level of nesting.

var_labels

character. Variable labels for 1 or more variables

cvar

character(1). The variable, if any, which the content function should accept. Defaults to NA.

table_names

character. Names for the tables representing each atomic analysis. Defaults to var.

topleft

character. Override values for the "top left" material to be displayed during printing.

align

character(1) or NULL. Alignment the value should be rendered with. It defaults to "center" if NULL is used. See rtables_aligns for currently supported alignments.

page_by

logical(1). Should pagination be forced between different children resulting form this split.

page_prefix

character(1). Prefix, to be appended with the split value, when forcing pagination between the children of this split/table

format_na_str

character(1). String which should be displayed when formatted if this cell's value(s) are all NA.

section_div

character(1). String which should be repeated as a section divider after each group defined by this split instruction, or NA_character_ (the default) for no section divider.

na_str

character(1). String that should be displayed when the value of x is missing. Defaults to "NA".

Value

NULL (this is an argument template dummy function)

See also