Pagination of a TableTree

Pagination of a TableTree

Usage

pag_tt_indices(
  tt,
  lpp = 15,
  min_siblings = 2,
  nosplitin = character(),
  colwidths = NULL,
  max_width = NULL,
  verbose = FALSE
)

paginate_table(
  tt,
  page_type = "letter",
  font_family = "Courier",
  font_size = 12,
  lineheight = 1,
  landscape = FALSE,
  pg_width = NULL,
  pg_height = NULL,
  margins = c(top = 0.5, bottom = 0.5, left = 0.75, right = 0.75),
  lpp,
  cpp,
  min_siblings = 2,
  nosplitin = character(),
  colwidths = NULL,
  tf_wrap = FALSE,
  max_width = NULL,
  verbose = FALSE
)

Arguments

tt

TableTree (or related class). A TableTree object representing a populated table.

lpp

numeric. Maximum lines per page including (re)printed header and context rows

min_siblings

numeric. Minimum sibling rows which must appear on either side of pagination row for a mid-subtable split to be valid. Defaults to 2.

nosplitin

character. List of names of sub-tables where page-breaks are not allowed, regardless of other considerations. Defaults to none.

colwidths

numeric vector. Column widths for use with vertical pagination.

max_width

integer(1), character(1) or NULL. Width that title and footer (including footnotes) materials should be word-wrapped to. If NULL, it is set to the current print width of the session (`getOption("width")`). If set to `"auto"`, the width of the table (plus any table inset) is used. Ignored completely if `tf_wrap` is `FALSE`.

verbose

logical(1). Should extra debugging messages be shown. Defaults to FALSE.

page_type

character(1). Name of a page type. See `page_types`. Ignored when `pg_width` and `pg_height` are set directly.

font_family

character(1). Name of a font family. An error will be thrown if the family named is not monospaced. Defaults to Courier.

font_size

numeric(1). Fontsize, defaults to 12.

lineheight

numeric(1). Lineheight, defaults to 1.

landscape

logical(1). Should the dimensions of `page_type` be inverted for landscape? Defaults to `FALSE`, ignored when `pg_width` and `pg_height` are set directly.

pg_width

numeric(1). Page width in inches.

pg_height

numeric(1). Page height in inches.

margins

numeric(4). Named numeric vector containing `'top'`, `'bottom'`, `'left'`, and `'right'` margins in inches. Defaults to `.5` inches for both vertical margins and `.75` for both horizontal margins.

cpp

numeric(1) or NULL. Width (in characters) of the pages for horizontal pagination. NULL (the default) indicates no horizontal pagination should be done.

tf_wrap

logical(1). Should the texts for title, subtitle, and footnotes be wrapped?

Value

for pag_tt_indices a list of paginated-groups of row-indices of tt. For paginate_table, The subtables defined by subsetting by the indices defined by pag_tt_indices.

Note

This is our first take on pagination. We will refine pagination in subsequent releases. Currently only pagination in the row space work. Pagination in the column space will be added in the future.

Examples

s_summary <- function(x) {
 if (is.numeric(x)) {
     in_rows(
         "n" = rcell(sum(!is.na(x)), format = "xx"),
         "Mean (sd)" = rcell(c(mean(x, na.rm = TRUE), sd(x, na.rm = TRUE)),
                             format = "xx.xx (xx.xx)"),
         "IQR" = rcell(IQR(x, na.rm = TRUE), format = "xx.xx"),
         "min - max" = rcell(range(x, na.rm = TRUE), format = "xx.xx - xx.xx")
     )
 } else if (is.factor(x)) {

     vs <- as.list(table(x))
     do.call(in_rows, lapply(vs, rcell, format = "xx"))

 } else (
     stop("type not supported")
 )
}

lyt <- basic_table() %>%
split_cols_by(var = "ARM") %>%
    analyze(c("AGE", "SEX", "BEP01FL", "BMRKR1", "BMRKR2", "COUNTRY"), afun = s_summary)

tbl <- build_table(lyt, ex_adsl)
tbl
#>                        A: Drug X      B: Placebo     C: Combination
#> ———————————————————————————————————————————————————————————————————
#> AGE                                                                
#>   n                       134             134             132      
#>   Mean (sd)          33.77 (6.55)    35.43 (7.90)     35.43 (7.72) 
#>   IQR                    11.00           10.00           10.00     
#>   min - max          21.00 - 50.00   21.00 - 62.00   20.00 - 69.00 
#> SEX                                                                
#>   F                       79              77               66      
#>   M                       51              55               60      
#>   U                        3               2               4       
#>   UNDIFFERENTIATED         1               0               2       
#> BEP01FL                                                            
#>   Y                       68              63               66      
#>   N                       66              71               66      
#> BMRKR1                                                             
#>   n                       134             134             132      
#>   Mean (sd)           5.97 (3.55)     5.70 (3.31)     5.62 (3.49)  
#>   IQR                    4.16            4.06             3.88     
#>   min - max          0.41 - 17.67    0.65 - 14.24     0.17 - 21.39 
#> BMRKR2                                                             
#>   LOW                     50              45               40      
#>   MEDIUM                  37              56               42      
#>   HIGH                    47              33               50      
#> COUNTRY                                                            
#>   CHN                     74              81               64      
#>   USA                     10              13               17      
#>   BRA                     13               7               10      
#>   PAK                     12               9               10      
#>   NGA                      8               7               11      
#>   RUS                      5               8               6       
#>   JPN                      5               4               9       
#>   GBR                      4               3               2       
#>   CAN                      3               2               3       
#>   CHE                      0               0               0       

nrow(tbl)
#> [1] 33

row_paths_summary(tbl)
#> rowname               node_class    path                                                           
#> ———————————————————————————————————————————————————————————————————————————————————————————————————
#> AGE                   LabelRow      ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, AGE                  
#>   n                   DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, AGE, n               
#>   Mean (sd)           DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, AGE, Mean (sd)       
#>   IQR                 DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, AGE, IQR             
#>   min - max           DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, AGE, min - max       
#> SEX                   LabelRow      ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, SEX                  
#>   F                   DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, SEX, F               
#>   M                   DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, SEX, M               
#>   U                   DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, SEX, U               
#>   UNDIFFERENTIATED    DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, SEX, UNDIFFERENTIATED
#> BEP01FL               LabelRow      ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, BEP01FL              
#>   Y                   DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, BEP01FL, Y           
#>   N                   DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, BEP01FL, N           
#> BMRKR1                LabelRow      ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, BMRKR1               
#>   n                   DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, BMRKR1, n            
#>   Mean (sd)           DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, BMRKR1, Mean (sd)    
#>   IQR                 DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, BMRKR1, IQR          
#>   min - max           DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, BMRKR1, min - max    
#> BMRKR2                LabelRow      ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, BMRKR2               
#>   LOW                 DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, BMRKR2, LOW          
#>   MEDIUM              DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, BMRKR2, MEDIUM       
#>   HIGH                DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, BMRKR2, HIGH         
#> COUNTRY               LabelRow      ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, COUNTRY              
#>   CHN                 DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, COUNTRY, CHN         
#>   USA                 DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, COUNTRY, USA         
#>   BRA                 DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, COUNTRY, BRA         
#>   PAK                 DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, COUNTRY, PAK         
#>   NGA                 DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, COUNTRY, NGA         
#>   RUS                 DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, COUNTRY, RUS         
#>   JPN                 DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, COUNTRY, JPN         
#>   GBR                 DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, COUNTRY, GBR         
#>   CAN                 DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, COUNTRY, CAN         
#>   CHE                 DataRow       ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY, COUNTRY, CHE         

tbls <- paginate_table(tbl)
mf <- matrix_form(tbl, indent_rownames = TRUE)
w_tbls <- propose_column_widths(mf) # so that we have the same column widths

tmp <- lapply(tbls, print, widths = w_tbls)
#> <S4 Type Object>
#> attr(,"content")
#>    A: Drug X   B: Placebo   C: Combination
#> ——————————————————————————————————————————
#> attr(,"page_title_prefix")
#> [1] NA
#> attr(,"children")
#> attr(,"children")$AGE
#>                 A: Drug X      B: Placebo     C: Combination
#> ————————————————————————————————————————————————————————————
#> AGE                                                         
#>   n                134             134             132      
#>   Mean (sd)   33.77 (6.55)    35.43 (7.90)     35.43 (7.72) 
#>   IQR             11.00           10.00           10.00     
#>   min - max   21.00 - 50.00   21.00 - 62.00   20.00 - 69.00 
#> 
#> attr(,"children")$SEX
#>                      A: Drug X   B: Placebo   C: Combination
#> ————————————————————————————————————————————————————————————
#> SEX                                                         
#>   F                     79           77             66      
#>   M                     51           55             60      
#>   U                      3           2              4       
#>   UNDIFFERENTIATED       1           0              2       
#> 
#> attr(,"children")$BEP01FL
#>           A: Drug X   B: Placebo   C: Combination
#> —————————————————————————————————————————————————
#> BEP01FL                                          
#>   Y          68           63             66      
#>   N          66           71             66      
#> 
#> attr(,"children")$BMRKR1
#>                A: Drug X      B: Placebo    C: Combination
#> ——————————————————————————————————————————————————————————
#> BMRKR1                                                    
#>   n               134            134             132      
#>   Mean (sd)   5.97 (3.55)    5.70 (3.31)     5.62 (3.49)  
#>   IQR             4.16           4.06            3.88     
#>   min - max   0.41 - 17.67   0.65 - 14.24    0.17 - 21.39 
#> 
#> attr(,"children")$BMRKR2
#>            A: Drug X   B: Placebo   C: Combination
#> ——————————————————————————————————————————————————
#> BMRKR2                                            
#>   LOW         50           45             40      
#>   MEDIUM      37           56             42      
#>   HIGH        47           33             50      
#> 
#> attr(,"children")$COUNTRY
#>           A: Drug X   B: Placebo   C: Combination
#> —————————————————————————————————————————————————
#> COUNTRY                                          
#>   CHN        74           81             64      
#>   USA        10           13             17      
#>   BRA        13           7              10      
#>   PAK        12           9              10      
#>   NGA         8           7              11      
#>   RUS         5           8              6       
#>   JPN         5           4              9       
#>   GBR         4           3              2       
#>   CAN         3           2              3       
#>   CHE         0           0              0       
#> 
#> attr(,"rowspans")
#> data frame with 0 columns and 0 rows
#> attr(,"labelrow")
#> [LabelRow indent_mod 0]:    
#> attr(,"page_titles")
#> character(0)
#> attr(,"horizontal_sep")
#> [1] "—"
#> attr(,"trailing_section_div")
#> [1] NA
#> attr(,"col_info")
#> An InstantiatedColumnInfo object
#> Columns:
#> A: Drug X (ARM)
#> B: Placebo (ARM)
#> C: Combination (ARM)
#> 
#> attr(,"format")
#> `\001NULL\001`
#> attr(,"na_str")
#> [1] NA
#> attr(,"indent_modifier")
#> [1] 0
#> attr(,"table_inset")
#> [1] 0
#> attr(,"level")
#> [1] 0
#> attr(,"name")
#> [1] "ma_AGE_SEX_BEP01FL_BMRKR1_BMRKR2_COUNTRY"
#> attr(,"main_title")
#> [1] ""
#> attr(,"subtitles")
#> character(0)
#> attr(,"main_footer")
#> character(0)
#> attr(,"provenance_footer")
#> character(0)
#> attr(,"class")
#> [1] "TableTree"
#> attr(,"class")attr(,"package")
#> [1] "rtables"

tmp <- lapply(tbls, function(tbli) {
  cat(toString(tbli, widths = w_tbls))
  cat("\n\n")
  cat("~~~~ PAGE BREAK ~~~~")
  cat("\n\n")
})
#>                        A: Drug X      B: Placebo     C: Combination
#> ———————————————————————————————————————————————————————————————————
#> AGE                                                                
#>   n                       134             134             132      
#>   Mean (sd)          33.77 (6.55)    35.43 (7.90)     35.43 (7.72) 
#>   IQR                    11.00           10.00           10.00     
#>   min - max          21.00 - 50.00   21.00 - 62.00   20.00 - 69.00 
#> SEX                                                                
#>   F                       79              77               66      
#>   M                       51              55               60      
#>   U                        3               2               4       
#>   UNDIFFERENTIATED         1               0               2       
#> BEP01FL                                                            
#>   Y                       68              63               66      
#>   N                       66              71               66      
#> BMRKR1                                                             
#>   n                       134             134             132      
#>   Mean (sd)           5.97 (3.55)     5.70 (3.31)     5.62 (3.49)  
#>   IQR                    4.16            4.06             3.88     
#>   min - max          0.41 - 17.67    0.65 - 14.24     0.17 - 21.39 
#> BMRKR2                                                             
#>   LOW                     50              45               40      
#>   MEDIUM                  37              56               42      
#>   HIGH                    47              33               50      
#> COUNTRY                                                            
#>   CHN                     74              81               64      
#>   USA                     10              13               17      
#>   BRA                     13               7               10      
#>   PAK                     12               9               10      
#>   NGA                      8               7               11      
#>   RUS                      5               8               6       
#>   JPN                      5               4               9       
#>   GBR                      4               3               2       
#>   CAN                      3               2               3       
#>   CHE                      0               0               0       
#> 
#> 
#> ~~~~ PAGE BREAK ~~~~
#>