Create a Reporting Table

Reporting tables allow multiple values per cell, cell formatting and merging cells. Currently an rtable can be converted to html and ascii.

rtable(col.names, format = NULL, ...)

Arguments

col.names
vector with column names

format
a valid format string or a format function for rcells. To get a list of all valid format strings use get_rcell_formats.

...
each element is an rrow object

Value

rtable object

Details

Rtable objects can be currently exported to text with toString.rtable and to html with as_html. In future we would plan to add the as_latex and as_gridGrob outputting function.

Note that the formats propagate to the rrow and rcell if these do not specify their own format.

See also

rrow, rcell

Examples

tbl <- rtable(
  col.names = c("Treatement\nN=100", "Comparison\nN=300"),
  format = "xx (xx.xx%)",
  rrow("A", c(104, .2), c(100, .4)),
  rrow("B", c(23, .4), c(43, .5)),
  rrow(),
  rrow("this is a very long section header"),
  rrow("estimate", rcell(55.23, "xx.xx", colspan = 2)),
  rrow("95% CI", indent = 1, rcell(c(44.8, 67.4), format = "(xx.x, xx.x)", colspan = 2))
)

tbl

#>                                           Treatement        Comparison
#>                                             N=100             N=300   
#> ----------------------------------------------------------------------
#> A                                         104 (20%)         100 (40%) 
#> B                                          23 (40%)          43 (50%) 
#> 
#> this is a very long section header
#> estimate                                             55.23            
#>   95% CI                                          (44.8, 67.4)        

row.names(tbl)

#> [1] "A"                                  "B"                                 
#> [3] ""                                   "this is a very long section header"
#> [5] "estimate"                           "95% CI"                            

names(tbl)

#> [1] "Treatement\nN=100" "Comparison\nN=300"

tbl[1,2]

#> [1] 100.0   0.4
#> attr(,"colspan")
#> [1] 1
#> attr(,"class")
#> [1] "rcell"
#> attr(,"format")
#> [1] "xx (xx.xx%)"

tbl[3,2]

#> NULL

tbl[5,1]

#> [1] 55.23
#> attr(,"format")
#> [1] "xx.xx"
#> attr(,"colspan")
#> [1] 2
#> attr(,"class")
#> [1] "rcell"

tbl[5,2]

#> [1] 55.23
#> attr(,"format")
#> [1] "xx.xx"
#> attr(,"colspan")
#> [1] 2
#> attr(,"class")
#> [1] "rcell"

dim(tbl)

#> [1] 6 2

nrow(tbl)

#> [1] 6

ncol(tbl)

#> [1] 2

tbl[[1]][[1]]

#> [1] 104.0   0.2
#> attr(,"colspan")
#> [1] 1
#> attr(,"class")
#> [1] "rcell"
#> attr(,"format")
#> [1] "xx (xx.xx%)"

as_html(tbl)

#> <table class="table table-condensed table-hover">
#>   <tr>
#>     <th></th>
#>     <th align="center" class="text-center">
#>       Treatement
#>       <br/>
#>       N=100
#>     </th>
#>     <th align="center" class="text-center">
#>       Comparison
#>       <br/>
#>       N=300
#>     </th>
#>   </tr>
#>   <tr>
#>     <td class="rowname" align="left">A</td>
#>     <td align="center">104 (20%)</td>
#>     <td align="center">100 (40%)</td>
#>   </tr>
#>   <tr>
#>     <td class="rowname" align="left">B</td>
#>     <td align="center">23 (40%)</td>
#>     <td align="center">43 (50%)</td>
#>   </tr>
#>   <tr>
#>     <td colspan="3" class="rowname" align="left"></td>
#>   </tr>
#>   <tr>
#>     <td colspan="3" class="rowname" align="left">this is a very long section header</td>
#>   </tr>
#>   <tr>
#>     <td class="rowname" align="left">estimate</td>
#>     <td colspan="2" align="center">55.23</td>
#>   </tr>
#>   <tr>
#>     <td class="rowname" align="left" style="padding-left: 3ch">95% CI</td>
#>     <td colspan="2" align="center">(44.8, 67.4)</td>
#>   </tr>
#> </table>

# NOT RUN {
Viewer(tbl)
# }

# colspans

tbl2 <- rtable(
  c("A", "B", "C", "D", "E"),
  format = "xx",
  rrow("r1", 1, 2, 3, 4, 5),
  rrow("r2", rcell("sp2", colspan = 2), "sp1", rcell("sp2-2", colspan = 2))
)

tbl2

#>            A          B          C          D          E 
#> ---------------------------------------------------------
#> r1         1          2          3          4          5 
#> r2             sp2              sp1            sp2-2     

# NOT RUN {
Viewer(tbl2)
# }

tbl2[1,3]

#> [1] 3
#> attr(,"colspan")
#> [1] 1
#> attr(,"class")
#> [1] "rcell"
#> attr(,"format")
#> [1] "xx"

tbl2[2,1]

#> [1] "sp2"
#> attr(,"colspan")
#> [1] 2
#> attr(,"class")
#> [1] "rcell"
#> attr(,"format")
#> [1] "xx"

tbl2[2,2]

#> [1] "sp2"
#> attr(,"colspan")
#> [1] 2
#> attr(,"class")
#> [1] "rcell"
#> attr(,"format")
#> [1] "xx"

tbl2[2,3]

#> [1] "sp1"
#> attr(,"colspan")
#> [1] 1
#> attr(,"class")
#> [1] "rcell"
#> attr(,"format")
#> [1] "xx"

tbl2[2,4]

#> [1] "sp2-2"
#> attr(,"colspan")
#> [1] 2
#> attr(,"class")
#> [1] "rcell"
#> attr(,"format")
#> [1] "xx"

tbl2[2,5]

#> [1] "sp2-2"
#> attr(,"colspan")
#> [1] 2
#> attr(,"class")
#> [1] "rcell"
#> attr(,"format")
#> [1] "xx"

# custom format
my_format <- function(x, output) {
   paste(x, collapse = "/")
}
tbl3 <- rtable(
  c("A", "B"),
  format = my_format,
  rrow("row1", c(1,2,3,4), letters[1:10])
)
tbl3

#>                      A                          B         
#> ----------------------------------------------------------
#> row1              1/2/3/4              a/b/c/d/e/f/g/h/i/j