Create a Reporting Table

Reporting tables allow multiple values per cell, cell formatting and merging cells. Currently an rtable can be converted to html and ascii.

rtable(header, ..., format = NULL)

Arguments

header
either a vector with column names or an object returned by rheader if special formating and multi-line headers are needed

...
each element is an rrow object

format
a valid format string or a format function for rcells. To get a list of all valid format strings use list_rcell_format_labels. If format is NULL then the elements of a cell get pasted separated by a comma.

Value

an rtable object

Details

Rtable objects can be currently exported to text with toString and to html with as_html. In future we would plan to add the as_latex and as_gridGrob outputting function.

Note that the formats propagate to the rrow and rcell if these do not specify their own format.

See also

Examples

# Table with multirow header
mtbl <- rtable(
  header = rheader(
    rrow(row.name = NULL, rcell("Sepal.Length", colspan = 2), rcell("Petal.Length", colspan=2)),
    rrow(NULL, "mean", "median", "mean", "median")
  ),
  rrow(
    row.name = "All Species",
    mean(iris$Sepal.Length), median(iris$Sepal.Length),
    mean(iris$Petal.Length), median(iris$Petal.Length),
    format = "xx.xx"
  )
)

mtbl

#>                        Sepal.Length                Petal.Length    
#>                     mean         median         mean         median
#> -------------------------------------------------------------------
#> All Species         5.84          5.8           3.76          4.35 

names(mtbl) # always first row of header

#> [1] "Sepal.Length" "Sepal.Length" "Petal.Length" "Petal.Length"

# Single row header

tbl <- rtable(
  header = c("Treatement\nN=100", "Comparison\nN=300"),
  format = "xx (xx.xx%)",
  rrow("A", c(104, .2), c(100, .4)),
  rrow("B", c(23, .4), c(43, .5)),
  rrow(),
  rrow("this is a very long section header"),
  rrow("estimate", rcell(55.23, "xx.xx", colspan = 2)),
  rrow("95% CI", indent = 1, rcell(c(44.8, 67.4), format = "(xx.x, xx.x)", colspan = 2))
)

tbl

#>                                           Treatement        Comparison
#>                                             N=100             N=300   
#> ----------------------------------------------------------------------
#> A                                         104 (20%)         100 (40%) 
#> B                                          23 (40%)          43 (50%) 
#> 
#> this is a very long section header
#> estimate                                             55.23            
#>   95% CI                                          (44.8, 67.4)        

row.names(tbl)

#> [1] "A"                                  "B"                                 
#> [3] ""                                   "this is a very long section header"
#> [5] "estimate"                           "95% CI"                            

names(tbl)

#> [1] "Treatement\nN=100" "Comparison\nN=300"

# Subsetting
tbl[1,2]

#> 100 (40%)

tbl[3,2]

#> NULL

tbl[5,1]

#> 55.23

tbl[5,2]

#> 55.23

tbl[1:3]

#>          Treatement        Comparison
#>            N=100             N=300   
#> -------------------------------------
#> A        104 (20%)         100 (40%) 
#> B         23 (40%)          43 (50%) 
#> 

# Data Structure methods
dim(tbl)

#> [1] 6 2

nrow(tbl)

#> [1] 6

ncol(tbl)

#> [1] 2

names(tbl)

#> [1] "Treatement\nN=100" "Comparison\nN=300"

# Output: html
as_html(tbl)

#> <table class="table table-condensed table-hover">
#>   <tr>
#>     <th class="rowname text-left"></th>
#>     <th class="text-center">
#>       Treatement
#>       <br/>
#>       N=100
#>     </th>
#>     <th class="text-center">
#>       Comparison
#>       <br/>
#>       N=300
#>     </th>
#>   </tr>
#>   <tr>
#>     <td class="rowname text-left">A</td>
#>     <td class="text-center">104 (20%)</td>
#>     <td class="text-center">100 (40%)</td>
#>   </tr>
#>   <tr>
#>     <td class="rowname text-left">B</td>
#>     <td class="text-center">23 (40%)</td>
#>     <td class="text-center">43 (50%)</td>
#>   </tr>
#>   <tr>
#>     <td class="rowname text-left" colspan="3"></td>
#>   </tr>
#>   <tr>
#>     <td class="rowname text-left" colspan="3">this is a very long section header</td>
#>   </tr>
#>   <tr>
#>     <td class="rowname text-left">estimate</td>
#>     <td colspan="2" class="text-center">55.23</td>
#>   </tr>
#>   <tr>
#>     <td class="rowname text-left" style="padding-left: 3ch">95% CI</td>
#>     <td colspan="2" class="text-center">(44.8, 67.4)</td>
#>   </tr>
#> </table>

# NOT RUN {
Viewer(tbl)
# }

# Colspans

tbl2 <- rtable(
  c("A", "B", "C", "D", "E"),
  format = "xx",
  rrow("r1", 1, 2, 3, 4, 5),
  rrow("r2", rcell("sp2", colspan = 2), "sp1", rcell("sp2-2", colspan = 2))
)

tbl2

#>            A          B          C          D          E 
#> ---------------------------------------------------------
#> r1         1          2          3          4          5 
#> r2             sp2              sp1            sp2-2     

# Custom format with functions (might be deprecated soon)
my_format <- function(x, output) {
   paste(x, collapse = "/")
}
tbl3 <- rtable(
  c("A", "B"),
  format = my_format,
  rrow("row1", c(1,2,3,4), letters[1:10])
)
tbl3

#>                      A                          B         
#> ----------------------------------------------------------
#> row1              1/2/3/4              a/b/c/d/e/f/g/h/i/j