Tabulate a numeric vector

By default each cell reports the mean based on the associated vector.

# S3 method for numeric
rtabulate(x, col_by = no_by("col_1"), FUN = mean,
  ..., format = NULL, row.name = NULL, indent = 0,
  col_wise_args = NULL)

Arguments

x
a vecor

col_by
a factor of length nrow(x) that defines which levels in col_by define a column. If data should not be split into columns use the no_by function.

FUN
a function that processes the cell data

...
arguments passed to FUN

format
if FUN does not return a formatted rcell then the format is applied

row.name
if NULL then the FUN argument is deparsed and used as row.name of the rrow

indent
non-negative integer where 0 means that the row should not be indented

col_wise_args
a named list containing collections (e.g. vectors or lists) with data elements for each column of the resulting table. The data elements are then passed to the named argument FUN corresponding to the element name of the outer list. Hence, the length and order of each collection must match the levels in col_by. See examples.

Value

an rtable object

Examples

rtabulate(iris$Sepal.Length)

#>                  col_1      
#> ----------------------------
#> mean        5.84333333333333

rtabulate(iris$Sepal.Length, col_by = no_by("Sepal.Length"))

#>               Sepal.Length  
#> ----------------------------
#> mean        5.84333333333333

with(iris,  rtabulate(x = Sepal.Length, col_by = Species, row.name = "mean"))

#>               setosa          versicolor        virginica 
#> ----------------------------------------------------------
#> mean          5.006             5.936             6.588   

SL <- iris$Sepal.Length
Sp <- iris$Species
rbind(
  rtabulate(SL, Sp, length, row.name = "n"),
  rtabulate(SL, Sp, function(x)c(mean(x), sd(x)), format = "xx.xx (xx.xx)", row.name = "Mean (SD)"),
  rtabulate(SL, Sp, median, row.name = "Median"),
  rtabulate(SL, Sp, range, format = "xx.xx - xx.xx", row.name = "Min - Max")
)

#>                    setosa           versicolor          virginica 
#> ------------------------------------------------------------------
#> n                    50                 50                 50     
#> Mean (SD)        5.01 (0.35)        5.94 (0.52)        6.59 (0.64)
#> Median                5                 5.9                6.5    
#> Min - Max         4.3 - 5.8           4.9 - 7           4.9 - 7.9 

x <- 1:100
cb <- factor(rep(LETTERS[1:3], c(20, 30, 50)))

rtabulate(
  x = x, col_by = cb, FUN = function(x, N) list(mean(x), sd(x), N),
  format = sprintf_format("%.2f (%.2f) and %i"), row.name = "Mean (SD) and N",
  col_wise_args = list(N = table(cb))
)

#>                                 A                           B                           C          
#> ---------------------------------------------------------------------------------------------------
#> Mean (SD) and N        10.50 (5.92) and 20         35.50 (8.80) and 30         75.50 (14.58) and 50