Split functions

Split functions

remove_split_levels(excl)

keep_split_levels(only, reorder = TRUE)

drop_split_levels(df, spl, vals = NULL, labels = NULL, trim = FALSE)

drop_and_remove_levels(excl)

reorder_split_levels(neworder, newlabels = neworder, drlevels = TRUE)

trim_levels_in_group(innervar)

Arguments

excl
character. Levels to be excluded (they will not be reflected in the resulting table structure regardless of presence in the data).

only
character. Levels to retain (all others will be dropped).

reorder
logical(1). Should the order of only be used as the order of the children of the split. defaults to TRUE

df
dataset (data.frame or tibble)

spl
A Split object defining a partitioning or analysis/tabulation of the data.

vals
ANY. For internal use only.

labels
character. Labels to use for the remaining levels instead of the existing ones.

trim
logical(1). Should splits corresponding with 0 observations be kept when tabulating.

neworder
character. New order or factor levels.

newlabels
character. Labels for (new order of) factor levels

drlevels
logical(1). Should levels in the data which do not appear in neworder be dropped. Defaults to TRUE

innervar
character(1). Variable whose factor levels should be trimmed (ie empty levels dropped) separately within each grouping defined at this point in the structure

Examples

l <- basic_table() %>%
  split_cols_by("ARM") %>%
  split_rows_by("COUNTRY", split_fun = remove_split_levels(c("USA", "CAN", "CHE", "BRA"))) %>%
  analyze("AGE")

build_table(l, DM)

#>          A: Drug X   B: Placebo   C: Combination
#> ------------------------------------------------
#> CHN                                             
#>   Mean     36.08       34.12          33.71     
#> PAK                                             
#>   Mean     35.38       33.12          36.75     
#> NGA                                             
#>   Mean     31.2         31.4          35.78     
#> RUS                                             
#>   Mean     33.33        34.2            33      
#> JPN                                             
#>   Mean     31.2         32.5           36.2     
#> GBR                                             
#>   Mean      32           29             30      

l <- basic_table() %>%
  split_cols_by("ARM") %>%
  split_rows_by("COUNTRY", split_fun = keep_split_levels(c("USA", "CAN", "BRA"))) %>%
  analyze("AGE")

build_table(l, DM)

#>          A: Drug X   B: Placebo   C: Combination
#> ------------------------------------------------
#> USA                                             
#>   Mean     36.77       32.57          36.41     
#> CAN                                             
#>   Mean      36           34            29.5     
#> BRA                                             
#>   Mean     31.78       30.62          36.14     

l <- basic_table() %>%
  split_cols_by("ARM") %>%
  split_rows_by("SEX", split_fun = drop_split_levels) %>%
  analyze("AGE")

build_table(l, DM)

#>          A: Drug X   B: Placebo   C: Combination
#> ------------------------------------------------
#> F                                               
#>   Mean     33.71       33.84          34.89     
#> M                                               
#>   Mean     36.55        32.1          34.28     

l <- basic_table() %>%
  split_cols_by("ARM") %>%
  split_rows_by("SEX", split_fun = drop_and_remove_levels(c("M", "U"))) %>%
  analyze("AGE")

build_table(l, DM)

#>          A: Drug X   B: Placebo   C: Combination
#> ------------------------------------------------
#> F                                               
#>   Mean     33.71       33.84          34.89