Create Enriched flat value table with paths

This function creates a flat tabular file of cell values and corresponding paths via path_enriched_df. I then writes that data.frame out as a tsv file.

Usage

export_as_tsv(
  tt,
  file = NULL,
  path_fun = collapse_path,
  value_fun = collapse_values
)

import_from_tsv(file)

Arguments

tt

TableTree (or related class). A TableTree object representing a populated table.

file

character(1). The path of the file to written to or read from.

path_fun

function. Function to transform paths into single-string row/column names.

value_fun

function. Function to transform cell values into cells of the data.frame. Defaults to collapse_values which creates strings where multi-valued cells are collapsed together, separated by |.

Value

NULL silently for export_as_tsv, a data.frame with re-constituted list values for export_as_tsv.

Details

By default (ie when value_func is not specified, List columns where at least one value has length > 1 are collapsed to character vectors by collapsing the list element with "|".

Note

There is currently no round-trip capability for this type of export. You can read values exported this way back in via import_from_tsv but you will receive only the data.frame version back, NOT a TableTree.