Validate and Assert valid table structure (Experimental).

Validate and Assert valid table structure (Experimental).

Usage

validate_table_struct(tt)

assert_valid_table(tt, warn_only = FALSE)

Arguments

tt

TableTree

A TableTree (rtables-built table) is considered degenerate if

it contains no subtables or data rows (content rows do not count)

it contains a subtable which is degenerate by the criterion above

validate_table_struct assesses whether tt has a valid (non-degenerate) structure.

assert_valid_table asserts a table must have a valid structure, and throws an informative error (the default) or warning (if warn_only is TRUE) if the table is degenerate (has invalid structure or contains one or more invalid substructures

warn_only

logical(1). Should a warning be thrown instead of an error? Defaults to FALSE

Value

for validate_table_struct a logical value indicating valid structure; assert_valid_table is called for its side-effect of throwing an error or warning for degenerate tables.

Note

This function is experimental and the exact text of the warning/error is subject to change in future releases.

Examples

validate_table_struct(rtable("hahaha"))
#> [1] FALSE
if (FALSE) assert_valid_table(rtable("oops"))